% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{add_percent_mito,Seurat-method}
\alias{add_percent_mito,Seurat-method}
\title{Annotate percent mitochondrial reads per cell}
\usage{
\S4method{add_percent_mito}{Seurat}(object, object_assay = "gene")
}
\arguments{
\item{object}{A object}

\item{object_assay}{gene}

\item{organism}{mouse}
}
\value{
a single cell obejct with cell metadata column containing mitochondrial percentage
}
\description{
Add a Read Count Categorical Variable to Seurat Object (based on nCount_RNA)
}
