% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{add_rc_cat_meta_to_seu}
\alias{add_rc_cat_meta_to_seu}
\title{Add Read Count Category}
\usage{
add_rc_cat_meta_to_seu(seu, thresh = 1e+05)
}
\arguments{
\item{thresh}{}
}
\description{
Add a Read Count Categorical Variable to Seurat Object (based on nCount_RNA)
}
