% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{add_read_count_col,Seurat-method}
\alias{add_read_count_col,Seurat-method}
\title{Annotate Low Read Count Category}
\usage{
\S4method{add_read_count_col}{Seurat}(object, thresh = 1e+05)
}
\arguments{
\item{object}{A object}

\item{thresh}{Set a threshold for low read count}
}
\value{
a single cell obejct with cell metadata column containing read counts
}
\description{
Add a Read Count Categorical Variable to Seurat Object (based on nCount_RNA)
}
