% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{add_read_count_col}
\alias{add_read_count_col}
\title{Annotate Low Read Count Category}
\usage{
add_read_count_col(seu, thresh = 1e+05)
}
\arguments{
\item{seu}{A seurat object}

\item{thresh}{Set a threshold for low read count}
}
\description{
Add a Read Count Categorical Variable to Seurat Object (based on nCount_RNA)
}
