% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{annotate_cell_cycle,Seurat-method}
\alias{annotate_cell_cycle,Seurat-method}
\title{Annotate Cell Cycle}
\usage{
\S4method{annotate_cell_cycle}{Seurat}(object, organism = "human")
}
\arguments{
\item{object}{A single cell object}

\item{organism}{organism. Human or mouse, default "human"}
}
\value{
a single cell object
}
\description{
Annotate Cell Cycle for Gene and Transcript Single Cell Objects
}
