% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{annotate_cell_cycle}
\alias{annotate_cell_cycle}
\title{Annotate Cell Cycle}
\usage{
annotate_cell_cycle(object)
}
\arguments{
\item{object}{A SingleCellExperiment object}
}
\value{
a SingleCellExperiment object
}
\description{
Annotate Cell Cycle for Gene and Transcript SingleCellExperiment Objects
}
