% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{annotate_excluded}
\alias{annotate_excluded}
\title{Annotate Exclusion Criteria}
\usage{
annotate_excluded(seu, ...)
}
\arguments{
\item{excluded_cells}{a named list of cells to be excluded of the form list(set1 = c(cell1, celll2), set2 = c(cell3, cell4))
all other cells will be marked NA in a column titled "excluded_because"}
}
\description{
Annotate Exclusion Criteria
}
