% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{arrange_ptime_heatmaps}
\alias{arrange_ptime_heatmaps}
\title{Arrange processed monocle CDS objects containing heatmaps}
\usage{
arrange_ptime_heatmaps(cds_list)
}
\arguments{
\item{cds_list}{}
}
\description{
Arrange processed monocle CDS objects containing heatmaps
}
