% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3.R
\name{assign_clusters_to_cds}
\alias{assign_clusters_to_cds}
\title{Assign Clusters to CDS}
\usage{
assign_clusters_to_cds(cds, clusters)
}
\arguments{
\item{clusters}{}
}
\description{
Assign Clusters to CDS
}
