% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{by_row_corr_between_pt_heatmaps}
\alias{by_row_corr_between_pt_heatmaps}
\title{calculate row-wise correlation between two dataframes}
\usage{
by_row_corr_between_pt_heatmaps(df1, df2)
}
\arguments{
\item{df2}{}
}
\description{
calculate row-wise correlation between two dataframes
}
