% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{calc_cor_across_heatmaps}
\alias{calc_cor_across_heatmaps}
\title{calculate correlation between heatmap matrices in processed cds_gene_list}
\usage{
calc_cor_across_heatmaps(cds_set, cds_set_name)
}
\arguments{
\item{cds_set_name}{}
}
\value{

}
\description{
calculate correlation between heatmap matrices in processed cds_gene_list
}
