% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc.genes.cyclone}
\alias{cc.genes.cyclone}
\title{Cyclone cell cycle pairs by symbol}
\format{
a list of dataframes with G1, G2, and S gene expression
\describe{
\item{G1}{G1 gene symbols}
\item{G2}{G2 gene symbols}
\item{S}{S gene symbols}
...
}
}
\source{
cyclone
}
\usage{
cc.genes.cyclone
}
\description{
cell cycle genes with paired expression represented by HGNC symbol
}
\keyword{datasets}
