% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype_markers}
\alias{celltype_markers}
\title{Retinal Cell Type Marker Genes}
\format{
A dataframe with 3 columns and 99 rows
\describe{
\item{gene}
\item{cell_type}
\item{species}
}
}
\source{
Cui et al. (2020). Transcriptomic Analysis of the Developmental Similarities and Differences Between the Native Retina and Retinal Organoids. Invest. Ophthalmol. Vis. Sci. 61, 6–6.
}
\usage{
celltype_markers
}
\description{
Retinal Cell Type Marker Genes
}
\keyword{datasets}
