% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{changeEmbedParams}
\alias{changeEmbedParams}
\title{Change Embedding Parameters}
\usage{
changeEmbedParams(input, output, session, object)
}
\arguments{
\item{object}{}
}
\description{
Change Embedding Parameters
}
