% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{changeEmbedParamsui}
\alias{changeEmbedParamsui}
\title{Change Embedding Parameters UI}
\usage{
changeEmbedParamsui(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Change Embedding Parameters UI
}
