% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevreul-package.R
\docType{package}
\name{chevreul-package}
\alias{chevreul}
\alias{chevreul-package}
\title{chevreul: Tools for Managing Seurat Objects}
\description{
Tools for using seurat with single cell projects
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whtns/chevreul}
  \item \url{https://whtns.github.io/chevreul/}
  \item Report bugs at \url{https://github.com/cobriniklab/chevreul/issues}
}

}
\author{
\strong{Maintainer}: Kevin Stachelek \email{kevin.stachelek@gmail.com}

Authors:
\itemize{
  \item Bhavana Bhat \email{bhavanabhat29@gmail.com}
}

Other contributors:
\itemize{
  \item Roman Hillje (Author of included Cerebro fragments) [contributor]
}

}
\keyword{internal}
