% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_cols,Seurat-method}
\alias{combine_cols,Seurat-method}
\title{Reformat object Metadata}
\usage{
\S4method{combine_cols}{Seurat}(object, cols, new_col)
}
\arguments{
\item{object}{A object}

\item{cols}{Columns}

\item{new_col}{New columns}
}
\value{
updated cell level metadata
}
\description{
Reformat object Metadata by Coalesced Columns
}
