% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_cols}
\alias{combine_cols}
\title{Reformat object Metadata}
\usage{
combine_cols(object, cols, new_col)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{cols}{Columns}

\item{new_col}{New columns}
}
\value{
updated cell level metadata
}
\description{
Reformat object Metadata by Coalesced Columns
}
