% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{combine_looms}
\alias{combine_looms}
\title{Combine Loom Files}
\usage{
combine_looms(projectPaths, newProjectPath)
}
\arguments{
\item{newProjectPath}{}
}
\value{

}
\description{
Combine Loom Files
}
