% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_human_object_to_mouse}
\alias{convert_human_object_to_mouse}
\title{Convert SingleCellExperiment Objects from Human to Mouse}
\usage{
convert_human_object_to_mouse(object, ...)
}
\arguments{
\item{object}{Human SingleCellExperiment object}

\item{...}{to be passed to \code{convert_symbols_by_species}}
}
\value{
a single cell object
}
\description{
Convert SingleCellExperiment Objects from Human to Mouse
}
