% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_human_seu_to_mouse}
\alias{convert_human_seu_to_mouse}
\title{Convert Seurat Objects from Human to Mouse}
\usage{
convert_human_seu_to_mouse(seu)
}
\arguments{
\item{seu}{}
}
\value{

}
\description{
Convert Seurat Objects from Human to Mouse
}
