% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{convert_monoclev2_to_objectv3}
\alias{convert_monoclev2_to_objectv3}
\title{Convert a Seurat V3 object to a Monocle v2 object}
\usage{
convert_monoclev2_to_objectv3(
  object,
  assay = "gene",
  slot = "data",
  return_census = FALSE,
  sig_slice = 1000
)
}
\arguments{
\item{object}{}
}
\description{
Convert a Seurat V3 object to a Monocle v2 object
}
