% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_mouse_object_to_human}
\alias{convert_mouse_object_to_human}
\title{Convert SingleCellExperiment Objects from Mouse to Human}
\usage{
convert_mouse_object_to_human(object)
}
\arguments{
\item{object}{Mouse object}
}
\value{
a single cell object
}
\description{
Convert SingleCellExperiment Objects from Mouse to Human
}
\examples{

convert_mouse_object_to_human(baron2016singlecell)
}
