% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_mouse_seu_to_human}
\alias{convert_mouse_seu_to_human}
\title{Convert Seurat Objects from Mouse to Human}
\usage{
convert_mouse_seu_to_human(seu)
}
\arguments{
\item{seu}{}
}
\description{
Convert Seurat Objects from Mouse to Human
}
