% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_object_list_to_multimodal}
\alias{convert_object_list_to_multimodal}
\title{convert object list to multimodal object}
\usage{
convert_object_list_to_multimodal(object_list)
}
\arguments{
\item{object_list}{}
}
\value{
a list of single cell objects
}
\description{
convert object list to multimodal object
}
