% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{convert_object_to_cds}
\alias{convert_object_to_cds}
\title{Convert a Seurat Object to a Monocle Cell Data Set}
\usage{
convert_object_to_cds(object, resolution = 1, min_expression = 0.05)
}
\arguments{
\item{object}{}
}
\description{
Convert a Seurat Object to a Monocle Cell Data Set
}
\examples{
processed_object <- clustering_workflow(human_gene_transcript_object)
cds <- convert_object_to_cds(processed_object)
}
