% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_seu_list_to_multimodal}
\alias{convert_seu_list_to_multimodal}
\title{convert seurat list to multimodal object}
\usage{
convert_seu_list_to_multimodal(seu_list)
}
\arguments{
\item{seu_list}{}
}
\description{
convert seurat list to multimodal object
}
