% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R, R/monocle3.R
\name{convert_seu_to_cds}
\alias{convert_seu_to_cds}
\title{Convert a Seurat Object to a Monocle Cell Data Set}
\usage{
convert_seu_to_cds(seu, resolution = 1)

convert_seu_to_cds(seu, resolution = 1)
}
\arguments{
\item{seu}{}

\item{seu}{}
}
\description{
Convert a Seurat Object to a Monocle Cell Data Set

Convert a Seurat Object to a Monocle Cell Data Set
}
