% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{convert_seuv3_to_monoclev2}
\alias{convert_seuv3_to_monoclev2}
\title{Convert a Seurat V3 object to a Monocle v2 object}
\usage{
convert_seuv3_to_monoclev2(seu)
}
\arguments{
\item{seu}{}
}
\description{
Convert a Seurat V3 object to a Monocle v2 object
}
