% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_symbols_by_biomart}
\alias{convert_symbols_by_biomart}
\title{convert gene symbols between species}
\usage{
convert_symbols_by_biomart(src_genes, src_species, host = "uswest.ensembl.org")
}
\arguments{
\item{src_genes}{a vector of hgnc/mgi symbols}

\item{src_species}{"mouse" or "human"}

\item{host}{the biomart host to use try also useast.ensembl.org}
}
\value{

}
\description{
convert gene symbols between species
}
