% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_symbols_by_species}
\alias{convert_symbols_by_species}
\title{convert gene symbols between species}
\usage{
convert_symbols_by_species(src_genes, src_species)
}
\arguments{
\item{src_species}{}
}
\description{
convert gene symbols between species
}
