% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_v3_to_v5}
\alias{convert_v3_to_v5}
\title{Convert seurat object to seurat V5 format}
\usage{
convert_v3_to_v5(seu_v3)
}
\arguments{
\item{seu_v3}{a seurat object}
}
\description{
Convert seurat object to seurat V5 format
}
\examples{
convert_v3_to_v5(human_gene_transcript_seu)
}
