% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{create_proj_db}
\alias{create_proj_db}
\title{Create Project database}
\usage{
create_proj_db(projects_dir = "/dataVolume/storage/single_cell_projects/")
}
\value{

}
\description{
Create Project database
}
