% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{create_proj_list}
\alias{create_proj_list}
\title{Create List of Projects}
\usage{

  create_proj_list(projects_dir = "/dataVolume/storage/single_cell_projects",
  sub_dirs = c("sc_cone_devel", "sc_RB_devel", "integrated_projects"))
}
\arguments{
\item{proj_matrices}{}
}
\description{
Create List of Projects
}
