% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{create_proj_matrix}
\alias{create_proj_matrix}
\title{Create a Table of single Cell Projects}
\usage{
create_proj_matrix(proj_list)
}
\arguments{
\item{proj_list}{List of projects}
}
\description{
Uses a list of projects to create a matrix of single cell projects
}
