% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{cross_check_heatmaps}
\alias{cross_check_heatmaps}
\title{Title}
\usage{
cross_check_heatmaps(
  monocle_list,
  query_name,
  set_row_order = F,
  cluster_rows = T,
  ...
)
}
\arguments{
\item{cluster_rows}{}
}
\value{

}
\description{
Title
}
