% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{cross_species_integrate}
\alias{cross_species_integrate}
\title{Integrate Seurat Objects from Mouse to Human}
\usage{
cross_species_integrate(mouse_object_list, human_object_list)
}
\arguments{
\item{mouse_object_list}{List of mouse single cell objects}

\item{human_object_list}{List of human single cell objects}
}
\value{
a single cell object
}
\description{
Integrate Seurat Objects from Mouse to Human
}
\examples{

cross_species_integrate(list(baron2016singlecell = baron2016singlecell), list(panc8 = panc8))

}
