% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_ui.R
\name{default_helper}
\alias{default_helper}
\title{Default Shiny Helper}
\usage{
default_helper(ui_element, title = "", content = "test", type = "inline", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Default Shiny Helper
}
