% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{diffex}
\alias{diffex}
\title{Differential Expression}
\usage{
diffex(
  input,
  output,
  session,
  seu,
  featureType,
  selected_cells,
  tests = c(`t-test` = "t", `wilcoxon rank-sum test` = "wilcox",
    `Likelihood-ratio test (bimodal)` = "bimod", MAST = "MAST")
)
}
\arguments{
\item{tests}{}
}
\description{
Differential Expression
}
