% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{displaySubset}
\alias{displaySubset}
\title{Display a Subset of Cells}
\usage{
displaySubset(input, output, session, seu, plot_types)
}
\arguments{
\item{plot_types}{}
}
\description{
Display a Subset of Cells
}
