% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_app.R
\name{dockerChevreulApp}
\alias{dockerChevreulApp}
\title{Create a obect app on a shinyproxy docker instance}
\usage{
dockerChevreulApp(
  object = NULL,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000,
  db_name = "single-cell-projects.db"
)
}
\arguments{
\item{object}{a obect object}

\item{appTitle}{a title for the app}

\item{organism_type}{human or mouse}

\item{futureMb}{the megabytes available for the future package}

\item{db_name}{path to sqlite database listing bigwig files for cells}
}
\value{
a dockerized shiny app
}
\description{
Create a obect app on a shinyproxy docker instance
}
