% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_app.R
\name{dockerSeuratApp}
\alias{dockerSeuratApp}
\title{Create a obect app on a shinyproxy docker instance}
\usage{
dockerSeuratApp(
  object = panc8,
  loom_path = NULL,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000,
  bigwig_db = "~/.cache/chevreul/bw-files.db"
)
}
\arguments{
\item{object}{a obect object}

\item{loom_path}{path to a loom file}

\item{futureMb}{#' @param bigwigb  path to sqlite database listing bigwig files for cells in the obect object}
}
\description{
Create a obect app on a shinyproxy docker instance
}
\examples{

/dontrun{
dockerSeuratApp(panc8)
}

}
