% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_app.R
\name{dockerSingleCellExperimentApp}
\alias{dockerSingleCellExperimentApp}
\title{Create a obect app on a shinyproxy docker instance}
\usage{
dockerSingleCellExperimentApp(
  object = NULL,
  loom_path = NULL,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000,
  bigwig_db = "~/.cache/chevreul/bw-files.db"
)
}
\arguments{
\item{object}{a obect object}

\item{loom_path}{path to a loom file}

\item{appTitle}{a title for the app}

\item{organism_type}{human or mouse}

\item{futureMb}{the megabytes available for the future package}

\item{bigwig_db}{path to sqlite database listing bigwig files for cells in the obect object}
}
\value{
a dockerized shiny app
}
\description{
Create a obect app on a shinyproxy docker instance
}
\examples{
\donttest{
dockerSingleCellExperimentApp(chevreul_sce)
}

}
