% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{filter_low_rc_cells}
\alias{filter_low_rc_cells}
\title{Filter our Cells from Seurat below read count threshold}
\usage{
filter_low_rc_cells(seu, read_thresh = 1e+05)
}
\arguments{
\item{seu}{A seurat object}

\item{read_thresh}{}
}
\description{
Filter our Cells from Seurat below read count threshold
}
