% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{filter_merged_object}
\alias{filter_merged_object}
\title{Filter a Single SingleCellExperiment Object}
\usage{
filter_merged_object(object, filter_var, filter_val, .drop = .drop)
}
\arguments{
\item{object}{A singlecell objects}

\item{filter_var}{filter variable}

\item{filter_val}{filter values}

\item{.drop}{whether to drop values}
}
\value{
a single cell object
}
\description{
Filter a Single SingleCellExperiment Object
}
