% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{filter_merged_objects}
\alias{filter_merged_objects}
\title{Filter a List of SingleCellExperiment Objects}
\usage{
filter_merged_objects(objects, filter_var, filter_val, .drop = F)
}
\arguments{
\item{objects}{single cell projects}

\item{filter_var}{filter variable}

\item{filter_val}{filter values}

\item{.drop}{whether to drop from single cell object}
}
\value{
a list of single cell objects
}
\description{
Filter SingleCellExperiment Objects by custom variable and reset experiment to uncorrected "gene"
}
