% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{filter_merged_seus}
\alias{filter_merged_seus}
\title{Filter Seurat Objects}
\usage{
filter_merged_seus(seus, filter_var, filter_val)
}
\arguments{
\item{tag}{}
}
\description{
Filter Seurat Objects by custom variable and reset assay to uncorrected "RNA"
}
