% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{find_all_markers,Seurat-method}
\alias{find_all_markers,Seurat-method}
\title{Find All Markers}
\usage{
\S4method{find_all_markers}{Seurat}(object, group_by = NULL, object_assay = "gene", ...)
}
\arguments{
\item{object}{An object.}

\item{group_by}{A metadata variable to group by.}

\item{object_assay}{Assay to use, Default "gene".}

\item{...}{extra args passed to stash_marker_features}
}
\value{
a single cell object containing marker genes
}
\description{
Find all markers at a range of resolutions
}
\examples{
markers_stashed_object <- find_all_markers(panc8)
marker_genes <- Misc(markers_stashed_object, "markers")
str(marker_genes)
}
