% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_reference_integrate.R
\name{find_markers}
\alias{find_markers}
\title{Find Cell Type Markers in a Seurat Object}
\usage{
find_markers(object, num_features)
}
\arguments{
\item{object}{A object}
}
\value{
a single cell object
}
\description{
Find Cell Type Markers in a Seurat Object
}
