% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_new_metadata,Seurat-method}
\alias{format_new_metadata,Seurat-method}
\title{Add object Metadata}
\usage{
\S4method{format_new_metadata}{Seurat}(object, datapath)
}
\arguments{
\item{object}{A object}

\item{datapath}{Path to file containing metadata}
}
\value{
a single cell object
}
\description{
Adds data to the object to produce a object with metadata added.
}
