% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_new_metadata}
\alias{format_new_metadata}
\title{Add object Metadata}
\usage{
format_new_metadata(object, datapath)
}
\arguments{
\item{object}{A object}

\item{datapath}{Path to file containing metadata}
}
\description{
Adds data to the object to produce a object with metadata added.
}
