% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{gene_test_plot}
\alias{gene_test_plot}
\title{Plot Gene expression over pseudotime on reference and query trajectories from cellAlign}
\usage{
gene_test_plot(expGlobalRef, expGlobalQuery, trajRef, trajQuery)
}
\arguments{
\item{trajQuery}{}
}
\description{
Plot Gene expression over pseudotime on reference and query trajectories from cellAlign
}
