% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{geneEnrichment}
\alias{geneEnrichment}
\title{Gene Enrichment}
\usage{
geneEnrichment(input, output, session, seu, diffex_results)
}
\arguments{
\item{diffex_clusters}{}
}
\value{

}
\description{
Gene Enrichment
}
