% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genes_from_object,Seurat-method}
\alias{genes_from_object,Seurat-method}
\title{Get genes from Object}
\usage{
\S4method{genes_from_object}{Seurat}(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
a vector of genes in a single cell object
}
\description{
Get genes from the object
}
