% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{genes_to_transcripts}
\alias{genes_to_transcripts}
\title{Gene Symbols to Ensembl Transcript Ids}
\usage{
genes_to_transcripts(genes, organism = "human")
}
\arguments{
\item{genes}{}
}
\value{

}
\description{
convert hgnc gene symbols to ensembl transcript ids
}
