% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_cell_metadata}
\alias{get_cell_metadata}
\title{Get cell metadata from a given object}
\usage{
get_cell_metadata(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
dataframe containing object metadata
}
\description{
Get cell metadata from a given object
}
\examples{


chevreul_sce <- mockSCE(ncells=200, ngenes=1000)
get_cell_metadata(chevreul_sce)
}
