% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_feature_types}
\alias{get_feature_types}
\title{Get Feature Types}
\usage{
get_feature_types(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
vector of feature types in an object
}
\description{
Get Feature Types
}
\examples{


chevreul_sce <- mockSCE(ncells=200, ngenes=1000)
get_feature_types(chevreul_sce)
}
