% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classes.R
\name{get_features,Seurat-method}
\alias{get_features,Seurat-method}
\title{Get feature names}
\usage{
\S4method{get_features}{Seurat}(object, ...)
}
\arguments{
\item{object}{a single cell object}

\item{...}{extra args passed to seurat class method}
}
\value{
variable features from a single cell object
}
\description{
Get feature names
}
