% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_features}
\alias{get_features}
\title{Get feature names}
\usage{
get_features(object, experiment = "gene")
}
\arguments{
\item{object}{a single cell object}

\item{experiment}{"gene" or "transcript"}
}
\value{
variable features from a single cell object
}
\description{
Get feature names
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
get_features(chevreul_sce)
}
