% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_child_exp_metadata.R
\name{get_meta}
\alias{get_meta}
\title{Get Metadata for a project}
\usage{
get_meta(proj_path)
}
\arguments{
\item{proj_path}{a project path}
}
\value{
a filepath to a project metadata file
}
\description{
Get Metadata for a project
}
