% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_object_metadata}
\alias{get_object_metadata}
\title{Get object metadata}
\usage{
get_object_metadata(object)
}
\arguments{
\item{object}{a SingleCellExperiment object}
}
\value{
variable features from a SingleCellExperiment object
}
\description{
Get object metadata
}
\examples{

data(small_example_dataset)
get_object_metadata(small_example_dataset)
}
