% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transcripts_from_object,Seurat-method}
\alias{get_transcripts_from_object,Seurat-method}
\title{Get Transcripts in object}
\usage{
\S4method{get_transcripts_from_object}{Seurat}(object, gene, organism = "human")
}
\arguments{
\item{object}{A object}

\item{gene}{Gene of intrest}

\item{organism}{Organism}
}
\value{
transcripts constituting a gene of interest in a single cell object
}
\description{
Get transcript ids in objects for one or more gene of interest
}
\examples{

RXRG_transcripts <- get_transcripts_from_object(human_gene_transcript_object, "RXRG")

}
