% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transcripts_from_object}
\alias{get_transcripts_from_object}
\title{Get Transcripts in object}
\usage{
get_transcripts_from_object(object, gene, organism = "human")
}
\arguments{
\item{object}{A object}

\item{gene}{Gene of intrest}

\item{organism}{Organism}
}
\description{
Get transcript ids in objects for one or more gene of interest
}
\examples{
RXRG_transcripts <- get_transcripts_from_object(human_gene_transcript_object, "RXRG")

}
