% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{global_align}
\alias{global_align}
\title{Global Alignment of two trajectories with cellAlign}
\usage{
global_align(exp_ref, exp_query, traj_ref, traj_query)
}
\arguments{
\item{traj_query}{}
}
\description{
Global Alignment of two trajectories with cellAlign
}
