% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{global_cellalign}
\alias{global_cellalign}
\title{Global Alignment of two trajectories with cellAlign}
\usage{
global_cellalign(exp_ref, exp_query, traj_ref, traj_query, numPts = 200)
}
\arguments{
\item{traj_query}{}
}
\description{
Global Alignment of two trajectories with cellAlign
}
