% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{harmony_integrate}
\alias{harmony_integrate}
\title{Integrate small datasets with harmony}
\usage{
harmony_integrate(seu_list)
}
\arguments{
\item{seu_list}{List of two or more seurat objects to integrate}
}
\description{
Integrate small datasets with harmony
}
