% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_to_mouse_homologs}
\alias{human_to_mouse_homologs}
\title{Gene Homologs Between Human and Mouse}
\format{
A data frame with 23188 rows and 2 columns
\describe{
\item{HGNC.symbol}{human gene symbols}
\item{MGI.symbol}{mouse gene symbols}
...
}
}
\source{
bioMart
}
\usage{
human_to_mouse_homologs
}
\description{
Homologs drawn from Biomart
}
\keyword{datasets}
