% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_scvelo}
\alias{install_scvelo}
\title{Install scvelo}
\usage{
install_scvelo(method = "auto", conda = "auto")
}
\arguments{
\item{method}{deterministic, stochastic, or dynamical}

\item{conda}{conda env}
}
\value{
package installed
}
\description{
Install scvelo
}
\examples{
\donttest{install_scvelo()}
}
