% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{integrateProj}
\alias{integrateProj}
\title{Integrate Projects Server Function}
\usage{
integrateProj(input, output, session, proj_matrices, object, proj_dir, con)
}
\arguments{
\item{session}{}
}
\description{
Integrate Projects Server Function
}
