% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{integration_workflow,SingleCellExperiment-method}
\alias{integration_workflow,SingleCellExperiment-method}
\title{Integration Workflow}
\usage{
\S4method{integration_workflow}{SingleCellExperiment}(
  batches,
  excluded_cells = NULL,
  resolution = seq(0.2, 2, by = 0.2),
  experiment_name = "default_experiment",
  organism = "human",
  ...
)
}
\arguments{
\item{batches}{objects for all batches provided as a list. If named, the resulting integrated object will be identified with corresponding values in 'batch' metadata}

\item{excluded_cells}{named list of cells to exclude}

\item{resolution}{value(s) to control the clustering resolution via \code{Seurat::FindMarkers}}

\item{experiment_name}{arbitrary name to identify experiment}

\item{organism}{either "human" or "mouse"}

\item{...}{}
}
\description{
Integrate multiple objects and save to file
}
