% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_integrate.R
\name{label_transfer}
\alias{label_transfer}
\title{Transfer Labels Between Seurat Objects}
\usage{
label_transfer(ref_object, query_object, ref_name = NULL, query_name = NULL)
}
\arguments{
\item{ref_object}{reference object}

\item{query_object}{query object}

\item{ref_name}{reference name}

\item{query_name}{query name}
}
\value{
a single cell object
}
\description{
Transfer Labels Between Seurat Objects
}
