% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classes.R
\name{list_plot_types,Seurat-method}
\alias{list_plot_types,Seurat-method}
\title{Collate list of variables to be plotted}
\usage{
\S4method{list_plot_types}{Seurat}(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
plot_types a list of category_vars or continuous_vars
}
\description{
Collate list of variables to be plotted
}
