% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_plot_types}
\alias{list_plot_types}
\title{Collate list of variables to be plotted}
\usage{
list_plot_types(seu)
}
\arguments{
\item{seu}{}
}
\value{
plot_types
}
\description{
Collate list of variables to be plotted
}
