% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{load_bigwigs}
\alias{load_bigwigs}
\title{Create a tibble of big file paths by sample}
\usage{
load_bigwigs(seu, proj_dir)
}
\arguments{
\item{proj_dir}{}
}
\value{

}
\description{
Create a tibble of big file paths by sample
}
