% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{load_counts_by_tximport}
\alias{load_counts_by_tximport}
<<<<<<< HEAD
\title{Load Sample Counts then run Tximport}
=======
\title{Load Sample Counts then run Tximpor}
>>>>>>> master
\usage{
load_counts_by_tximport(
  proj_dir,
  type = "salmon",
  countsFromAbundance = "scaledTPM",
  edb = EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86
)
}
\arguments{
\item{edb}{}
}
\value{

}
\description{
<<<<<<< HEAD
Load Sample Counts then run Tximport
=======
Load Sample Counts then run Tximpor
>>>>>>> master
}
