% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{load_meta}
\alias{load_meta}
\title{Load Sample Metadata for a given project}
\usage{
load_meta(proj_dir)
}
\arguments{
\item{proj_dir}{}
}
\value{

}
\description{
Load Sample Metadata for a given project
}
