% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{load_object_from_proj,Seurat-method}
\alias{load_object_from_proj,Seurat-method}
\title{Load Seurat Files from a signle project path}
\usage{
\S4method{load_object_from_proj}{Seurat}(proj_dir, ...)
}
\arguments{
\item{proj_dir}{project directory}

\item{...}{extra args passed to load_object_path}
}
\value{
a single cell object
}
\description{
Load Seurat Files from a signle project path
}
