% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{load_object_path}
\alias{load_object_path}
\title{Read in Gene and Transcript SingleCellExperiment Objects}
\usage{
load_object_path(proj_dir = getwd(), prefix = "unfiltered")
}
\arguments{
\item{proj_dir}{path to project directory}

\item{prefix}{default "unfiltered"}
}
\value{
a single cell object
}
\description{
Read in Gene and Transcript SingleCellExperiment Objects
}
