% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{load_seurat_from_proj}
\alias{load_seurat_from_proj}
\title{Load Seurat Files from a signle project path}
\usage{
load_seurat_from_proj(proj_dir, features = "gene", ...)
}
\arguments{
\item{...}{}
}
\description{
Load Seurat Files from a signle project path
}
