% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{load_seurat_from_rds}
\alias{load_seurat_from_rds}
\title{Load Seurat files from a vector of project paths}
\usage{
load_seurat_from_rds(proj_dirs)
}
\arguments{
\item{proj_dirs}{}
}
\description{
Load Seurat files from a vector of project paths
}
