% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{load_seurat_path}
\alias{load_seurat_path}
\title{Read in Gene and Transcript Seurat Objects}
\usage{
load_seurat_path(proj_dir = getwd(), features = "gene", suffix = "")
}
\arguments{
\item{proj_dir}{}
}
\description{
Read in Gene and Transcript Seurat Objects
}
