% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{loadData}
\alias{loadData}
\title{Load Data}
\usage{
loadData(input, output, session, proj_dir, feature_types = c("gene"),
  selected_feature)
}
\arguments{
\item{proj_dir}{}
}
\description{
Load Data
}
