% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{local_align}
\alias{local_align}
\title{Align two trajectories via cellalign}
\usage{
local_align(exp_ref, exp_query, traj_ref, traj_query)
}
\arguments{
\item{traj_query}{}
}
\description{
Align two trajectories via cellalign
}
