% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{local_cellalign}
\alias{local_cellalign}
\title{Align two trajectories via cellalign}
\usage{
local_cellalign(
  exp_ref,
  exp_query,
  traj_ref,
  traj_query,
  Thresh = 0.2,
  winSz = 0.1,
  numPts = 200
)
}
\arguments{
\item{exp_ref}{reference experiment}

\item{exp_query}{query experiment}

\item{traj_ref}{reference trajectory}

\item{traj_query}{query trajectory}

\item{Thresh}{threhold}

\item{winSz}{window size}

\item{numPts}{number of points}
}
\value{
a list containing an alignment and an alignment plot
}
\description{
cell align
}
