% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_bigwig_db}
\alias{make_bigwig_db}
\title{Make Bigwig Database}
\usage{
make_bigwig_db(
  new_project = NULL,
  cache_location = "~/.cache/chevreul/",
  sqlite_db = "bw-files.db"
)
}
\arguments{
\item{sqlite_db}{}
}
\description{
Make Bigwig Database
}
