% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_chevreul_clean_names}
\alias{make_chevreul_clean_names}
\title{Clean Vector of Chevreul Names}
\usage{
make_chevreul_clean_names(myvec)
}
\arguments{
\item{myvec}{A vector of object names}
}
\value{
a clean vector of object names
}
\description{
Cleans names of objects provided in a vector form
}
