% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_complex_heatmap,Seurat-method}
\alias{make_complex_heatmap,Seurat-method}
\title{Plot Annotated Complexheatmap from Seurat object}
\usage{
\S4method{make_complex_heatmap}{Seurat}(
  object,
  features = NULL,
  group.by = "ident",
  cells = NULL,
  layer = "scale.data",
  assay = NULL,
  group.bar.height = 0.01,
  column_split = NULL,
  col_arrangement = "ward.D2",
  mm_col_dend = 30,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{Vector of features to plot. Features can come}

\item{group.by}{Name of one or more metadata columns to annotate columns by (for example, orig.ident)}

\item{cells}{Cells to retain}

\item{layer}{"counts" for raw data "scale.data" for log-normalized data}

\item{assay}{assay to display}

\item{group.bar.height}{height for group bars}

\item{column_split}{whether to split columns by metadat value}

\item{col_arrangement}{how to arrange columns whether with a dendrogram (Ward.D2, average, etc.) or exclusively by metadata category}

\item{mm_col_dend}{height of column dendrogram}

\item{...}{additional arguments passed to ComplexHeatmap::Heatmap}
}
\value{
a complexheatmap
}
\description{
Plot Annotated Complexheatmap from Seurat object
}
\examples{
\dontrun{top_50_features <- get_variable_features(human_gene_transcript_object)[1:50]}
\dontrun{make_complex_heatmap(human_gene_transcript_object, features = top_50_features)}
}
