% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{merge_small_objects}
\alias{merge_small_objects}
\title{Merge Small SingleCellExeriment Objects}
\usage{
merge_small_objects(..., k.filter = 50)
}
\arguments{
\item{k.filter}{minimum cell number for integration}

\item{object_list}{List of two or more singlecell objects}
}
\value{
a single cell object
}
\description{
Merge Small SingleCellExeriment Objects
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
merge_small_objects(
"small_batch1" = chevreul_sce[,1:40],
"small_batch2" = chevreul_sce[,41:80],
"large_batch" = chevreul_sce[,81:300])

}
