% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{merge_small_seus}
\alias{merge_small_seus}
\title{Merge Small Seurat Objects}
\usage{
merge_small_seus(seu_list, k.filter = 50)
}
\arguments{
\item{seu_list}{List of two or more seurat objects}
}
\description{
Merge Small Seurat Objects
}
