% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metadata_from_object,Seurat-method}
\alias{metadata_from_object,Seurat-method}
\title{Get metadata from object}
\usage{
\S4method{metadata_from_object}{Seurat}(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
a tibble with metadata from a single cell object
}
\description{
Get metadata from the given object
}
