% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_sce_app.R
\name{minimalSceApp}
\alias{minimalSceApp}
\title{Create a minimal object app}
\usage{
minimalSceApp(
  input_object = panc8,
  loom_path = NULL,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000,
  bigwig_db = "~/.cache/chevreul/bw-files.db"
)
}
\arguments{
\item{loom_path}{path to a loom file}

\item{organism_type}{Organism}

\item{futureMb}{}

\item{object}{a singlecell object}
}
\description{
Create a minimal object app
}
\examples{

/dontrun{
minimalSceApp(panc8)
}

}
