% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_features}
\alias{mito_features}
\title{Mitochondial Features}
\format{
a list of human  with G1, G2, and S gene expression
\describe{
\item{human}{human mitochondrial features: genes and transcripts}
\item{mouse}{mouse mitochondrial features: genes and transcripts}
...
}
}
\source{
bioMart
}
\usage{
mito_features
}
\description{
Mitochondrial genes in mouse and human
}
\keyword{datasets}
