% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_add_percent_mito}
\alias{mod_add_percent_mito}
\title{Annotate Low Read Count Category}
\usage{
mod_add_percent_mito(seu, assay, organism)
}
\arguments{
\item{seu}{A seurat object}

\item{organism}{}
}
\value{

}
\description{
Add a Read Count Categorical Variable to Seurat Object (based on nCount_RNA)
}
