% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_annotate_cell_cycle}
\alias{mod_annotate_cell_cycle}
\title{Annotate Cell Cycle}
\usage{
mod_annotate_cell_cycle(seu, assay, organism = "human", ...)
}
\arguments{
\item{seu_list}{A list of seurat objects}
}
\value{

}
\description{
Annotate Cell Cycle for Gene and Transcript Seurat Objects
}
