% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_clustering_workflow}
\alias{mod_clustering_workflow}
\title{Clustering Workflow}
\usage{
mod_clustering_workflow(
  seu = NULL,
  excluded_cells,
  resolution = seq(0.2, 2, by = 0.2),
  organism = "human",
  experiment_name = "default_experiment",
  ...
)
}
\arguments{
\item{excluded_cells}{named list of cells to exclude}

\item{resolution}{resolution(s) to use for clustering cells}

\item{...}{}

\item{feature_seus}{list of seurat objedevelcts named according to feature of interest ("gene" or "transcript")}
}
\value{

}
\description{
Cluster and Reduce Dimensions of a seurat object
}
\examples{
clustered_seu <- clustering_workflow(seurat_pancrease_reduced)
}
