% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_seu_from_tximport}
\alias{mod_seu_from_tximport}
\title{create a multimodal seurat object from genes and transcript counts from tximport}
\usage{
mod_seu_from_tximport(txi, meta_tbl, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
create a multimodal seurat object from genes and transcript counts from tximport
}
