% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_seurat_cluster}
\alias{mod_seurat_cluster}
\title{Run Louvain Clustering at Multiple Resolutions}
\usage{
mod_seurat_cluster(
  seu = seu,
  assays = NULL,
  resolution = 0.6,
  custom_clust = NULL,
  reduction = "pca",
  algorithm = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Run Louvain Clustering at Multiple Resolutions
}
