% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_seurat_integrate}
\alias{mod_seurat_integrate}
\title{Batch Correct Multiple Seurat Objects}
\usage{
mod_seurat_integrate(seu_list, method = "cca", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Batch Correct Multiple Seurat Objects
}
