% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_seurat_pipeline}
\alias{mod_seurat_pipeline}
\title{Run Seurat Pipeline}
\usage{
mod_seurat_pipeline(
  seu,
  feature = "none",
  resolution = 0.6,
  reduction = "pca",
  annotate_cell_cycle = TRUE,
  annotate_percent_mito = TRUE,
  organism = "human",
  ...
)
}
\arguments{
\item{resolution}{}
}
\value{

}
\description{
Preprocess, Cluster and Reduce Dimensions for a single seurat object
}
