% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_seurat_preprocess}
\alias{mod_seurat_preprocess}
\title{Preprocess Seurat Object}
\usage{
mod_seurat_preprocess(
  seu,
  scale = TRUE,
  normalize = TRUE,
  features = NULL,
  legacy_settings = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Preprocess Seurat Object
}
