% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_seurat_reduce_dimensions}
\alias{mod_seurat_reduce_dimensions}
\title{Dimensional Reduction}
\usage{
mod_seurat_reduce_dimensions(seu, assays = NULL, legacy_settings = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Run PCA, TSNE and UMAP on a seurat object
perplexity should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation
}
