% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_stash_marker_features}
\alias{mod_stash_marker_features}
\title{Stash Marker Genes in a Seurat Object}
\usage{
mod_stash_marker_features(metavar, seu, assay, top_n = 200)
}
\arguments{
\item{top_n}{}
}
\value{

}
\description{
Marker Genes will be stored in slot \verb{@misc$markers}
}
