% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSeurat_functions.R
\name{mod_update_seuratTools_object}
\alias{mod_update_seuratTools_object}
\title{Title}
\usage{
mod_update_seuratTools_object(
  seu_list,
  feature,
  resolution = seq(0.2, 2, by = 0.2),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
