% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{monocle}
\alias{monocle}
\title{Monocle Server Module}
\usage{
monocle(
  input,
  output,
  session,
  seu,
  plot_types,
  featureType,
  organism_type,
  reductions
)
}
\arguments{
\item{resolution}{}
}
\description{
Monocle Server Module
}
