% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{monocle_module_heatmap}
\alias{monocle_module_heatmap}
\title{Title}
\usage{
monocle_module_heatmap(
  cds,
  pr_deg_ids,
  seu_resolution,
  collapse_rows = TRUE,
  collapse_cols = TRUE,
  resolution = 10^seq(-6, -1)
)
}
\arguments{
\item{seu_resolution}{}
}
\value{

}
\description{
Title
}
