% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{object_cluster}
\alias{object_cluster}
\title{Run Louvain Clustering at Multiple Resolutions}
\usage{
object_cluster(
  object = object,
  resolution = 0.6,
  ustom_clust = NULL,
  reduction = "PCA",
  algorithm = 1,
  ...
)
}
\arguments{
\item{object}{A SingleCellExperiment objects}

\item{resolution}{Clustering resolution}

\item{reduction}{Set dimensional reduction object}

\item{algorithm}{1}

\item{...}{extra args passed to single cell packages}

\item{custom_clust}{custom cluster}
}
\value{
a SingleCellExperiment object with louvain clusters
}
\description{
Run Louvain Clustering at Multiple Resolutions
}
