% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_reference_integrate.R
\name{object_find_markers}
\alias{object_find_markers}
\title{Find Cell Type Markers in a Seurat Object}
\usage{
object_find_markers(object, num_features)
}
\arguments{
\item{object}{A object}
}
\description{
Find Cell Type Markers in a Seurat Object
}
