% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{object_from_tibbles}
\alias{object_from_tibbles}
\title{Create a Seurat Object from a set of tibbles}
\usage{
object_from_tibbles(exp_tbl, feature, meta_tbl)
}
\arguments{
\item{exp_tbl}{experiment tibble}

\item{feature}{feature type}

\item{meta_tbl}{metadata tibble}
}
\value{
a single cell object
}
\description{
Create a Seurat Object from a set of tibbles
}
