% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{object_integrate}
\alias{object_integrate}
\title{Batch Correct Multiple Single Cell Objects}
\usage{
object_integrate(object_list, method = "cca", organism = "human", ...)
}
\arguments{
\item{object_list}{List of two or more single cell objects}

\item{method}{Default "cca"}

\item{organism}{human or mouse}

\item{...}{extra args passed to object_reduce_dimensions}
}
\value{
an integrated single cell object
}
\description{
Batch Correct Multiple Single Cell Objects
}
