% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{object_integration_pipeline,SingleCellExperiment-method}
\alias{object_integration_pipeline,SingleCellExperiment-method}
\title{Run Seurat Integration}
\usage{
\S4method{object_integration_pipeline}{SingleCellExperiment}(
  object_list,
  resolution = seq(0.2, 2, by = 0.2),
  suffix = "",
  algorithm = 1,
  organism = "human",
  annotate_cell_cycle = FALSE,
  annotate_percent_mito = FALSE,
  reduction = "pca",
  ...
)
}
\arguments{
\item{object_list}{List of objects to be integrated}

\item{resolution}{Range of resolution}

\item{suffix}{a suffix to be appended to a file save in output dir}

\item{algorithm}{Algorithm for modularity optimization. Default 1:original Louvain algorithm}

\item{organism}{Default "human"}

\item{...}{}
}
\description{
Run batch correction, followed by:
\enumerate{
\item stashing of batches in metadata 'batch'
\item clustering with resolution 0.2 to 2.0 in increments of 0.2
\item saving to <proj_dir>/output/sce/\if{html}{\out{<feature>}}\emph{object}\if{html}{\out{<suffix>}}.rds
}
}
\examples{

batches <- panc8 \%>\%
  Seurat::SplitObject(split.by = "tech")

integrated_object <- object_integration_pipeline(batches)
}
