% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{object_integration_pipeline}
\alias{object_integration_pipeline}
\title{Run SingleCellExperiment Integration}
\usage{
object_integration_pipeline(
  object_list,
  resolution = seq(0.2, 2, by = 0.2),
  suffix = "",
  organism = "human",
  annotate_cell_cycle = FALSE,
  annotate_percent_mito = FALSE,
  reduction = "PCA",
  ...
)
}
\arguments{
\item{object_list}{List of objects to be integrated}

\item{resolution}{Range of resolution}

\item{suffix}{a suffix to be appended to a file save in output dir}

\item{organism}{Default "human"}

\item{annotate_cell_cycle}{whether to score cell cycle phases}

\item{annotate_percent_mito}{logical scalar whether to annotate mitochondrial percentage}

\item{reduction}{pca, umap, or tsne}

\item{...}{extra args passed to object_integrate}
}
\value{
an integrated SingleCellExperiment object
}
\description{
Run batch correction, followed by:
\enumerate{
\item stashing of batches in metadata 'batch'
\item clustering with resolution 0.2 to 2.0 in increments of 0.2
\item saving to <proj_dir>/output/sce/\if{html}{\out{<feature>}}\emph{object}\if{html}{\out{<suffix>}}.rds
}
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
batches <- splitByCol(chevreul_sce, "batch")
integrated_object <- object_integration_pipeline(batches)
}
