% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{object_pipeline,Seurat-method}
\alias{object_pipeline,Seurat-method}
\title{Run procdessing pipeline}
\usage{
\S4method{object_pipeline}{Seurat}(
  object,
  assay = "gene",
  resolution = 0.6,
  reduction = "pca",
  organism = "human",
  ...
)
}
\arguments{
\item{object}{A single cell object}

\item{assay}{Assay of interest in single cell object}

\item{resolution}{Resolution for clustering cells. Default set to 0.6.}

\item{reduction}{Dimensional reduction object}

\item{organism}{Organism}

\item{...}{extra parameters passed to internal functions}
}
\value{
a processed single cell object
}
\description{
This functions allows you to Preprocess, Cluster and Reduce Dimensions for a single object.
}
\examples{

processed_object <- object_pipeline(panc8)

}
