% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{object_reduce_dimensions,Seurat-method}
\alias{object_reduce_dimensions,Seurat-method}
\title{Dimensional Reduction}
\usage{
\S4method{object_reduce_dimensions}{Seurat}(
  object,
  assay = "gene",
  reduction = "pca",
  legacy_settings = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Assay of interest to be run on the singlecell objects}

\item{reduction}{Set dimensional reduction object}

\item{legacy_settings}{Use legacy settings}

\item{...}{Extra parameters passed to object_reduce_dimensions}
}
\value{
a single cell object with embeddings
}
\description{
Run PCA, TSNE and UMAP on a singlecell objects
perplexity should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation
}
