% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pad_sample_files}
\alias{pad_sample_files}
\title{pad sample numbers to prep for armor}
\usage{
pad_sample_files(proj_dir)
}
\arguments{
\item{proj_dir}{project directory}
}
\value{
a list of fastq files
}
\description{
pad sample numbers to prep for armor
}
