% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{pathwayEnrichment}
\alias{pathwayEnrichment}
\title{pathway enrichment}
\usage{
pathwayEnrichment(input, output, session, object, featureType)
}
\arguments{
\item{session}{}
}
\description{
pathway enrichment
}
