% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{plot_all_ptimes}
\alias{plot_all_ptimes}
\title{Plot a Set of Heatmaps Based on a provided pseudotime}
\usage{
plot_all_ptimes(monocle_list, query_name, sig_slice = 1000, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot a Set of Heatmaps Based on a provided pseudotime
}
