% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{plot_cds}
\alias{plot_cds}
\title{Plot a Monocle Cell Data Set}
\usage{
plot_cds(cds, color_cells_by = NULL, genes = NULL, return_plotly = TRUE)
}
\arguments{
\item{color_cells_by}{}
}
\description{
Plot a Monocle Cell Data Set
}
