% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_cycle_distribution,Seurat-method}
\alias{plot_cell_cycle_distribution,Seurat-method}
\title{Plot cell cycle distribution grouped by metadata}
\usage{
\S4method{plot_cell_cycle_distribution}{Seurat}(object)
}
\arguments{
\item{object}{A single cell object}
}
\value{
a ggplot of cell cycle scores
}
\description{
Plot ridge plots of G1, S, and G2M phases grouped by provided metadata
}
