% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dend_by_metadata}
\alias{plot_dend_by_metadata}
\title{Plot a dendrogram colored by Seurat metadata}
\usage{
plot_dend_by_metadata(seu, metavar = "batch", hclust_method = "ward.D2")
}
\arguments{
\item{metavar}{metadata variable from seurat object}

\item{hclust_method}{hierarchical clustering method as defined in hclust}
}
\value{

}
\description{
Plot a dendrogram colored by Seurat metadata
}
