% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{plot_feature_in_ref_query_ptime}
\alias{plot_feature_in_ref_query_ptime}
\title{Plot Expression of a Given Feature of a set of Pobjectdotimes}
\usage{
plot_feature_in_ref_query_ptime(
  cds_list,
  selected_cds,
  features = c("RXRG"),
  color_by = "State",
  relative_expr = FALSE,
  min_expr = 0.5,
  trend_df = 3,
  ...
)
}
\arguments{
\item{min_expr}{}
}
\description{
Plot Expression of a Given Feature of a set of Pobjectdotimes
}
