% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_markers}
\alias{plot_markers}
\title{Plot Cluster Marker Genes}
\usage{
plot_markers(
  seu,
  metavar,
  num_markers = 5,
  selected_values = NULL,
  return_plotly = TRUE,
  marker_method = c("presto", "genesorteR"),
  featureType = "gene",
  hide_pseudo = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot a dot plot of n marker features grouped by cell metadata
available methods are wilcoxon rank-sum test implemented in
\href{https://github.com/immunogenomics/presto}{presto} and specificity scores implemented in \href{https://github.com/mahmoudibrahim/genesorteR}{genesorteR}
}
\examples{

plot_markers(seurat_pancreas_reduced$gene, "tech", return_plot = TRUE)

}
