% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{plot_monocle_all_ptimes}
\alias{plot_monocle_all_ptimes}
\title{Plot a heatmap of differentially expressed genes over a query and set of reference pseudotimes}
\usage{
plot_monocle_all_ptimes(monocle_list, query_name, ...)
}
\arguments{
\item{query_name}{}
}
\description{
Plot a heatmap of differentially expressed genes over a query and set of reference pseudotimes
}
