% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{plot_monocle_features}
\alias{plot_monocle_features}
\title{Plot feature expression on a Monocle Cell Data Set}
\usage{
plot_monocle_features(cds, resolution, genes = NULL, ...)
}
\arguments{
\item{resolution}{}
}
\value{

}
\description{
Plot feature expression on a Monocle Cell Data Set
}
