% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_multiple_branches_heatmap}
\alias{plot_multiple_branches_heatmap}
\title{Plot pseudotime over multiple branches}
\usage{
plot_multiple_branches_heatmap(
  cds,
  branches,
  branches_name = NULL,
  cluster_rows = TRUE,
  hclust_method = "ward.D2",
  num_clusters = 6,
  hmcols = NULL,
  add_annotation_row = NULL,
  add_annotation_col = NULL,
  show_rownames = FALSE,
  use_gene_short_name = TRUE,
  norm_method = c("vstExprs", "log"),
  scale_max = 3,
  scale_min = -3,
  trend_formula = "~sm.ns(Pseudotime, df=3)",
  return_heatmap = FALSE,
  cores = 1
)
}
\arguments{
\item{cores}{}
}
\value{

}
\description{
Plot pseudotime over multiple branches
}
