% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3.R
\name{plot_pseudotime}
\alias{plot_pseudotime}
\title{Plot a Monocle Cell Data Set}
\usage{
plot_pseudotime(cds, resolution, color_cells_by = "louvain_cluster")
}
\arguments{
\item{color_cells_by}{}
}
\description{
Plot a Monocle Cell Data Set
}
