% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_readcount}
\alias{plot_readcount}
\title{Plot Read Count}
\usage{
plot_readcount(
  seu,
  metavar = "nCount_RNA",
  color.by = "batch",
  yscale = "linear",
  return_plotly = TRUE,
  ...
)
}
\arguments{
\item{return_plotly}{}
}
\value{

}
\description{
Plot Read Count
}
\examples{

plot_readcount(seurat_pancreas_reduced$gene)

plot_readcount(seurat_pancreas_reduced$gene, return_plotly = FALSE)

}
