% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_transcript_composition}
\alias{plot_transcript_composition}
\title{Plot Transcript Composition}
\usage{
plot_transcript_composition(
  seu_transcript,
  seu_gene,
  gene_symbol,
  group.by = "batch",
  standardize = FALSE,
  drop_zero = FALSE
)
}
\arguments{
\item{standardize}{}
}
\value{

}
\description{
plot the proportion of reads of a given gene map to each transcript
}
