% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{plot_velocity}
\alias{plot_velocity}
\title{Plot RNA velocity Computed by Velocyto.R}
\usage{
plot_velocity(vel, emb, cell.colors, format = "arrow")
}
\arguments{
\item{vel}{velocity from velocyto.R}

\item{emb}{embedding -- either tnse or umap}

\item{cell.colors}{A vector of colors to use for the plot}

\item{format}{either an arrow or grid format}
}
\description{
Plot RNA velocity Computed by Velocyto.R
}
