% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_velocity_arrows}
\alias{plot_velocity_arrows}
\title{Plot RNA velocity Computed by Velocyto.R}
\usage{
plot_velocity_arrows(
  seu,
  velocity,
  reduction = "umap",
  cell.colors,
  plot_format = "grid",
  arrow.scale = 3,
  ...
)
}
\arguments{
\item{velocity}{}
}
\value{

}
\description{
Plot RNA velocity Computed by Velocyto.R
}
