% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_violin}
\alias{plot_violin}
\title{plot Violin plot}
\usage{
plot_violin(
  seu,
  plot_var = "batch",
  plot_vals = NULL,
  features = "RXRG",
  return_plotly = TRUE,
  assay = "RNA",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plot Violin plot
}
