% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotCoverage}
\alias{plotCoverage}
\title{Title}
\usage{
plotCoverage(
  input,
  output,
  session,
  seu,
  plot_types,
  proj_dir,
  organism_type = "human"
)
}
\arguments{
\item{organism_type}{}
}
\value{

}
\description{
Title
}
