% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotDimRedui}
\alias{plotDimRedui}
\title{Plot Dimensionally Reduced Data UI}
\usage{
plotDimRedui(id, plot_types)
}
\arguments{
\item{plot_types}{}
}
\description{
Plot Dimensionally Reduced Data UI
}
