% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(input, output, session, seu, featureType, organism_type)
}
\arguments{
\item{organism_type}{}
}
\value{

}
\description{
Plot Heatmap
}
