% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotly_settings}
\alias{plotly_settings}
\title{Plotly settings}
\usage{
plotly_settings(plotly_plot, width = 600, height = 700)
}
\arguments{
\item{plotly_plot}{A plotly plot}

\item{width}{Default set to '600'}

\item{height}{Default set to '700'}
}
\description{
Change settings of a plotly plot
}
