% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotVelocity}
\alias{plotVelocity}
\title{RNA Velocity Server Module}
\usage{
plotVelocity(input, output, session, object, loom_path)
}
\arguments{
\item{loom_path}{}
}
\description{
RNA Velocity Server Module
}
