% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot Violin Server}
\usage{
plotViolin(input, output, session, object, featureType, organism_type)
}
\arguments{
\item{object}{Seurat object}

\item{featureType}{Gene or Transcript}

\item{organism_type}{Organism}
}
\description{
Plots a Violin plot of a single data (gene expression, metrics, etc.) in the server Seurat app.
}
