% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prep_armor_meta}
\alias{prep_armor_meta}
\title{prep armor metadata file}
\usage{
prep_armor_meta(proj_dir)
}
\arguments{
\item{proj_dir}{}
}
\value{
a path to a tsv
}
\description{
prep armor metadata file
}
