% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{prep_cell_metadata}
\alias{prep_cell_metadata}
\title{Prep Cell Metadata for Wiggleplotr}
\usage{
prep_cell_metadata(cell_metadata_file, var_of_interest, bigwig_dir)
}
\arguments{
\item{bigwig_dir}{}
}
\value{

}
\description{
Prep Cell Metadata for Wiggleplotr
}
