% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_plot_genes_in_pseudotime}
\alias{prep_plot_genes_in_pseudotime}
\title{Title}
\usage{
prep_plot_genes_in_pseudotime(cds, mygenes, resolution, partition = FALSE)
}
\arguments{
\item{resolution}{}
}
\value{

}
\description{
Title
}
