% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{prep_scvelo}
\alias{prep_scvelo}
\title{scvelo_assay}
\usage{
prep_scvelo(
  seu,
  loom_path,
  velocity_mode = c("deterministic", "stochastic", "dynamical"),
  ...
)
}
\arguments{
\item{seu}{a seurat object}

\item{loom_path}{path to matching loom file}

\item{...}{}

\item{group.by}{metadata to color plot}

\item{plot_method}{plotting method to use from scvelo}
}
\value{

}
\description{
run scvelo on a gene or transcript level seurat object
}
\examples{



}
