% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{process_monocle_child}
\alias{process_monocle_child}
\title{Preprocess a Monocle v2 object for heatmap based on provided pseudotime}
\usage{
process_monocle_child(
  ptime,
  monocle_cds,
  trend_formula = "~sm.ns(Pseudotime, df=3)"
)
}
\arguments{
\item{monocle_cds}{}
}
\description{
Preprocess a Monocle v2 object for heatmap based on provided pseudotime
}
