% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propagate_spreadsheet_changes}
\alias{propagate_spreadsheet_changes}
\title{Propagate Metadata Changes}
\usage{
propagate_spreadsheet_changes(changes)
}
\arguments{
\item{changes}{}
}
\value{

}
\description{
Propagate Metadata Changes
}
