% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{pull_gene_trx_seus}
\alias{pull_gene_trx_seus}
\title{Read in Gene and Transcript Seurat Objects}
\usage{
pull_gene_trx_seus(proj_dir)
}
\arguments{
\item{proj_dir}{}
}
\description{
Read in Gene and Transcript Seurat Objects
}
