% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classes.R
\name{pull_metadata,Seurat-method}
\alias{pull_metadata,Seurat-method}
\title{Pull the metadata from a given object}
\usage{
\S4method{pull_metadata}{Seurat}(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
dataframe containing object metadata
}
\description{
Pull the metadata from a given object
}
