% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_project_db}
\alias{read_project_db}
\title{Update a database of chevreul projects}
\usage{
read_project_db(
  projects_dir = NULL,
  destdir = "~/.cache/chevreul",
  destfile = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Update a database of chevreul projects
}
