% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{record_experiment_data,Seurat-method}
\alias{record_experiment_data,Seurat-method}
\title{Record Experiment Metadata}
\usage{
\S4method{record_experiment_data}{Seurat}(
  object,
  experiment_name = "default_experiment",
  organism = "human"
)
}
\arguments{
\item{object}{A object}

\item{experiment_name}{name of the experiment}

\item{organism}{human or mouse}
}
\value{
a single cell object
}
\description{
Records miscellaneous data
}
