% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{record_experiment_data}
\alias{record_experiment_data}
\title{Record Experiment Metadata}
\usage{
record_experiment_data(
  object,
  experiment_name = "default_experiment",
  organism = "human"
)
}
\arguments{
\item{object}{A object objet}

\item{organism}{}
}
\description{
Record Experiment Metadata
}
\examples{
logged_object <- record_experiment_data(human_gene_transcript_object, experiment_name = "human_gene_transcript", organism = "mouse")
Misc(logged_object, "experiment")
}
