% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_integrate.R
\name{reference_integrate}
\alias{reference_integrate}
\title{Transfer labels between gene or transcript objects}
\usage{
reference_integrate(
  ref_object,
  query_object,
  query_name = "fetal",
  ref_name = "organoid"
)
}
\arguments{
\item{ref_object}{reference object}

\item{query_object}{query object}

\item{query_name}{query name}

\item{ref_name}{reference name}
}
\value{
a single cell object
}
\description{
Transfer labels between gene or transcript objects
}
