% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{reformatMetadata}
\alias{reformatMetadata}
\title{Reformat Seurat Object Metadata Server}
\usage{
reformatMetadata(input, output, session, seu, featureType = "gene")
}
\arguments{
\item{seu}{}
}
\value{

}
\description{
Reformat Seurat Object Metadata Server
}
\examples{

}
