% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_edit.R
\name{reformatMetadataDR}
\alias{reformatMetadataDR}
\title{Reformat Seurat Object Metadata Server}
\usage{
reformatMetadataDR(
  input,
  output,
  session,
  object,
  featureType = "gene",
  col_bind = NULL,
  col_edit = TRUE,
  col_options = NULL,
  col_stretch = FALSE,
  col_names = TRUE,
  col_readonly = NULL,
  col_factor = FALSE,
  row_bind = NULL,
  row_edit = TRUE,
  save_as = NULL,
  title = NULL,
  logo = NULL,
  logo_size = 30,
  logo_side = "left",
  viewer = "dialog",
  viewer_height = 800,
  viewer_width = 2000,
  theme = "yeti",
  read_fun = "read.csv",
  read_args = NULL,
  write_fun = "write.csv",
  write_args = NULL,
  quiet = FALSE,
  code = FALSE,
  hide = FALSE
)
}
\arguments{
\item{object}{}
}
\description{
Reformat Seurat Object Metadata Server
}
