% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_edit.R
\name{reformatMetadataDRui}
\alias{reformatMetadataDRui}
\title{Reformat SingleCellExeriment Object Metadata UI}
\usage{
reformatMetadataDRui(id)
}
\arguments{
\item{id}{id}
}
\value{
UI to reformat the metadata of a single cell object
}
\description{
Reformat SingleCellExeriment Object Metadata UI
}
