% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{reformatMetadataui}
\alias{reformatMetadataui}
\title{Reformat Seurat Object Metadata UI}
\usage{
reformatMetadataui(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Reformat Seurat Object Metadata UI
}
