% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_features.R
\name{regress_by_features,Seurat-method}
\alias{regress_by_features,Seurat-method}
\title{Regress Seurat Object by Given Set of Genes}
\usage{
\S4method{regress_by_features}{Seurat}(object, feature_set, set_name, regress = TRUE)
}
\arguments{
\item{object}{A object}

\item{feature_set}{a set of features}

\item{set_name}{as a string}

\item{regress}{whether to regress}
}
\value{
a single cell object with features regressed
}
\description{
Regress Seurat Object by Given Set of Genes
}
\examples{

regressed_object <- regress_by_features(panc8, feature_set = cc.genes$s.genes, set_name = "s_genes")

}
