% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_features.R
\name{regress_by_features}
\alias{regress_by_features}
\title{Regress Seurat Object by Given Set of Genes}
\usage{
regress_by_features(object, feature_set, set_name, regress = TRUE, ...)
}
\arguments{
\item{object}{A object}

\item{set_name}{as a string}

\item{regress}{}
}
\description{
Regress Seurat Object by Given Set of Genes
}
\examples{

regressed_object <- regress_by_features(panc8, feature_set  = cc.genes$s.genes, set_name = "s_genes")

}
