% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_features.R
\name{regress_by_gene_set}
\alias{regress_by_gene_set}
\title{Regress Seurat Object by Given Set of Genes}
\usage{
regress_by_gene_set(seu_list, gene_set, set_name, ...)
}
\arguments{
\item{gene_set}{as a length 1 list containing a character vector of gene symbols}

\item{set_name}{as a string}

\item{...}{}
}
\description{
Regress Seurat Object by Given Set of Genes
}
