% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{reintegrate_object}
\alias{reintegrate_object}
\title{Reintegrate (filtered) singlecell objectss}
\usage{
reintegrate_object(object, suffix = "", reduction = "PCA", ...)
}
\arguments{
\item{object}{A singlecell objects}

\item{suffix}{to be appended to file saved in output dir}

\item{reduction}{to use default is pca}

\item{...}{extra args passed to object_integration_pipeline}
}
\value{
a single cell object
}
\description{
This function takes a SCE object and perfroms teh below steps
\enumerate{
\item split by batch
\item integrate
\item run integration pipeline and save
}
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
reintegrate_object(chevreul_sce)
}
