% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{reintegrate_seus}
\alias{reintegrate_seus}
\title{Reintegrate (filtered) seurat objects}
\usage{
reintegrate_seus(seus, suffix = "", reduction = "pca", ...)
}
\arguments{
\item{seus}{}
}
\description{
\enumerate{
\item split by batch
\item integrate
\item run integration pipeline and save
}
}
