% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rename_from_x_notation}
\alias{rename_from_x_notation}
\title{Rename cell ids from annoying old notation}
\usage{
rename_from_x_notation(cell_ids, batch_id)
}
\arguments{
\item{cell_ids}{cell ids}

\item{batch_id}{bath id}
}
\value{
a vector of cell ids
}
\description{
Rename cell ids from annoying old notation
}
