% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{rename_markers}
\alias{rename_markers}
\title{Convenience Function to transition existing seus}
\usage{
rename_markers(seu)
}
\arguments{
\item{seu}{}
}
\value{

}
\description{
Convenience Function to transition existing seus
}
