% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{rename_object,Seurat-method}
\alias{rename_object,Seurat-method}
\title{Give a new project name to a single cell object}
\usage{
\S4method{rename_object}{Seurat}(object, new_name)
}
\arguments{
\item{object}{A Seurat object}

\item{new_name}{New name to assign}
}
\value{
a renamed single cell object
}
\description{
Give a new project name to a single cell object
}
