% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reorg_object_files}
\alias{reorg_object_files}
\title{Reorganize objects to a multimodal format}
\usage{
reorg_object_files(projects_db = "~/.cache/chevreul/single-cell-projects.db")
}
\arguments{
\item{proj_dir}{}
}
\value{
a list of single cell objects
}
\description{
Reorganize objects to a multimodal format
}
