% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorg_seurat_files}
\alias{reorg_seurat_files}
\title{Reorganize seurat objects to a multimodal format}
\usage{
reorg_seurat_files(
  projects_db = "~/.cache/seuratTools/single-cell-projects.db"
)
}
\arguments{
\item{proj_dir}{}
}
\value{

}
\description{
Reorganize seurat objects to a multimodal format
}
