% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{replace_markers_in_proj}
\alias{replace_markers_in_proj}
\title{Convenience function to replace markers in seurat objects}
\usage{
replace_markers_in_proj(suffixes, features, proj_dir)
}
\arguments{
\item{proj_dir}{}
}
\description{
Convenience function to replace markers in seurat objects
}
