% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_object_metadata}
\alias{replace_object_metadata}
\title{Replace object Metadata}
\usage{
replace_object_metadata(object, datapath)
}
\arguments{
\item{object}{A object}

\item{datapath}{Path to csv file containing metadata}
}
\value{
a single cell object
}
\description{
Replace object metadata.
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
new_meta <- system.file("extdata", "new_meta.csv", package="chevreul")
replace_object_metadata(chevreul_sce, new_meta)
}
