% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seurat_from_child_metadata.R
\name{reset_object_meta}
\alias{reset_object_meta}
\title{Reset Seurat Metadata}
\usage{
reset_object_meta(object, new_meta)
}
\arguments{
\item{object}{A object}

\item{new_meta}{new object}
}
\value{
a single cell object with new metadata
}
\description{
Reset Seurat Metadata
}
