% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{rnaVelocity}
\alias{rnaVelocity}
\title{RNA Velocity}
\usage{
rnaVelocity(input, output, session, seu, featureType, format = "grid")
}
\arguments{
\item{format}{}
}
\value{

}
\description{
RNA Velocity
}
