% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{run_BEAM}
\alias{run_BEAM}
\title{Run BEAM from Monocle2}
\usage{
run_BEAM(
  HSMM,
  branch_point,
  branches,
  pt_param,
  pt_paramval,
  colorval,
  top_genes
)
}
\arguments{
\item{top_genes}{}
}
\value{

}
\description{
http://cole-trapnell-lab.github.io/monocle-release/docs/#differential-expression-analysis
}
