% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{run_census}
\alias{run_census}
\title{Run Census}
\usage{
run_census(sce, census_output_file)
}
\arguments{
\item{sce}{a single cell experiment object}

\item{census_output_file}{name to be given to output file}
}
\description{
Run Census
}
