% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{run_enrichmentbrowser}
\alias{run_enrichmentbrowser}
\title{Run Enrichment Browser on Differentially Expressed Genes}
\usage{
run_enrichmentbrowser(
  object,
  cluster_list,
  de_results,
  enrichment_method = c("ora"),
  ...
)
}
\arguments{
\item{object}{a single cell object}

\item{cluster_list}{a list of clusters}

\item{de_results}{differential expression results}

\item{enrichment_method}{enrichment method}

\item{...}{extra arguments passed to ggplot}
}
\value{
a list of enrichmentbrowser output
}
\description{
Run Enrichment Browser on Differentially Expressed Genes
}
