% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevreul_app.R
\name{run_object_de}
\alias{run_object_de}
\title{Run Differential Expression}
\usage{
run_object_de(
  object,
  cluster1,
  cluster2,
  resolution,
  diffex_scheme = "louvain",
  featureType,
  tests = c("t", "wilcox", "bimod")
)
}
\arguments{
\item{object}{single cell object}

\item{cluster1}{cluster 1}

\item{cluster2}{cluster 2}

\item{resolution}{resolution}

\item{diffex_scheme}{scheme for differential expression}

\item{featureType}{gene or transcript}

\item{tests}{t wicox or bimod}
}
\value{
a dataframe with differential expression information
}
\description{
Run Differential Expression
}
