% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_velocity.R
\name{run_rna_velocity}
\alias{run_rna_velocity}
\title{Run RNA Velocity}
\usage{
run_rna_velocity(seu, loom_path)
}
\arguments{
\item{seu}{Seurat object}

\item{loom_path}{path to a loom file generated by \href{https://velocyto.org/}{velocyto}}
}
\description{
Run RNA Velocity
}
