% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{run_scvelo}
\alias{run_scvelo}
\title{scvelo_assay}
\usage{
run_scvelo(
  seu,
  loom_path,
  assay = "gene",
  fit.quantile = 0.05,
  check_loom = FALSE,
  ...
)
}
\arguments{
\item{seu}{a seurat object}

\item{loom_path}{path to matching loom file}

\item{...}{}
}
\description{
run scvelo on a gene or transcript level seurat object
}
