% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{run_seurat_de}
\alias{run_seurat_de}
\title{Run Seurat Differential Expression}
\usage{
run_seurat_de(seu, cluster1, cluster2, resolution,
  diffex_scheme = "seurat")
}
\arguments{
\item{diffex_scheme}{}
}
\description{
Run Seurat Differential Expression
}
