% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_object}
\alias{save_object}
\title{Save object to \if{html}{\out{<project>}}/output/sce/\if{html}{\out{<feature>}}_object.rds}
\usage{
save_object(object, prefix = "unfiltered", proj_dir = getwd())
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{prefix}{a prefix for saving}

\item{proj_dir}{path to a project directory}
}
\value{
a path to an rds file containing a SingleCellExperiment object
}
\description{
Save object to \if{html}{\out{<project>}}/output/sce/\if{html}{\out{<feature>}}_object.rds
}
