% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{save_proj_feature_seus}
\alias{save_proj_feature_seus}
\title{Convenience function to save all seurat objects of a given project by feature and suffix}
\usage{
save_proj_feature_seus(seu_list, proj_dir)
}
\arguments{
\item{seu_list}{list of names seurat objects by suffix then feature; can generate with load_seurat_from_proj and purrr}

\item{proj_dir}{}
}
\value{

}
\description{
Convenience function to save all seurat objects of a given project by feature and suffix
}
