% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{scdeTest}
\alias{scdeTest}
\title{Compare two sets of cells with scde package}
\usage{
scdeTest(data.use, cells.1, cells.2, verbose = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Compare two sets of cells with scde package
}
