% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{sce_from_tibbles}
\alias{sce_from_tibbles}
\title{Create Single Cell Experiment from Tibbles}
\usage{
sce_from_tibbles(counts, colData, experimentdata = NULL)
}
\arguments{
\item{experimentdata}{}
}
\value{

}
\description{
Create Single Cell Experiment from Tibbles
}
