% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{sce_integrate}
\alias{sce_integrate}
\title{Batch Correct Multiple SingleCellExperiments}
\usage{
sce_integrate(object_list, method = "cca", organism = "human", ...)
}
\arguments{
\item{object_list}{List of two or more singlecell objects}

\item{method}{Default "cca"}

\item{organism}{human or mouse}

\item{...}{extra args passed to object_reduce_dimensions}
}
\value{
an integrated single cell object
}
\description{
Batch Correct Multiple SingleCellExperiments
}
