% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{set_colnames_txi}
\alias{set_colnames_txi}
\title{Set Column Names from \code{tximport}}
\usage{
set_colnames_txi(txi, colnames)
}
\arguments{
\item{colnames}{}
}
\value{

}
\description{
Set Column Names from \code{tximport}
}
