% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{set_feature_types}
\alias{set_feature_types}
\title{Set Feature Types}
\usage{
set_feature_types(object, feature_type)
}
\arguments{
\item{object}{a single cell object}

\item{feature_type}{feature type}
}
\value{
an object with assigned feature type
}
\description{
Set Feature Types
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce
set_feature_types(chevreul_sce, "transcript")
}
