% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classes.R
\name{set_metadata,Seurat-method}
\alias{set_metadata,Seurat-method}
\title{Add new metadata to the object metadata}
\usage{
\S4method{set_metadata}{Seurat}(object, meta.data)
}
\arguments{
\item{object}{a single cell object}

\item{meta.data}{new metadata}
}
\value{
object metadata
}
\description{
Add new metadata to the object metadata
}
