% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{set_metadata}
\alias{set_metadata}
\title{Add new metadata to the object metadata}
\usage{
set_metadata(object, meta.data)
}
\arguments{
\item{object}{a single cell object}

\item{meta.data}{new metadata}
}
\value{
object metadata
}
\description{
Add new metadata to the object metadata
}
