% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{SetDefaultAssay}
\alias{SetDefaultAssay}
\title{Reset default assay}
\usage{
SetDefaultAssay(object, new_assay)
}
\arguments{
\item{object}{A singlecell objects}

\item{new_assay}{Assay to set as default assay}
}
\description{
Function to reset the default assay of a singlecell objects
}
