% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seu_calcn}
\alias{seu_calcn}
\title{recalculate counts/features per cell for a seurat object}
\usage{
seu_calcn(seu, assay = "RNA", slot = "counts")
}
\arguments{
\item{slot}{}
}
\value{

}
\description{
recalculate counts/features per cell for a seurat object
}
