% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{seu_from_tibbles}
\alias{seu_from_tibbles}
\title{Create a Seurat Object from tximport and sample data}
\usage{
seu_from_tibbles(txi, colData, census_counts = NULL)
}
\arguments{
\item{census_counts}{}
}
\description{
Create a Seurat Object from tximport and sample data
}
