% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{seu_from_tximport}
\alias{seu_from_tximport}
\title{Create a Seurat Object from return of tximport}
\usage{
seu_from_tximport(txi, meta_tbl, ...)
}
\arguments{
\item{txi}{output from load_counts_from_stringtie}

\item{meta_tbl}{a tibble of cell metadata with cell ids as the first column}

\item{...}{}
}
\value{

}
\description{
Create a Seurat Object from return of tximport
}
