% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Seurat-class}
\alias{Seurat-class}
\alias{multiSeurat}
\title{The multiSeurat Class}
\description{
The multiSeurat object a multimodal seurat object with a minimum set of assays including gene-level expression (gene) and
transcript level expression (transcript) as well as assiociated embeddings and transformations. Just like a Seurat object it stores all information
associated with the dataset, including data, annotations, analyes, etc. To construct a multiSeurat object one must
provide expression matrices at the gene level (rows are genes, columns are cells), and at the transcript level (rows are transcripts, columns are cells)
}
\details{
Each Seurat object has a number of slots which store information. Key slots to access
are listed below.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of assays for this project}

\item{\code{meta.data}}{Contains meta-information about each cell, starting with number of genes detected (nGene)
and the original identity class (orig.ident); more information is added using \code{AddMetaData}}

\item{\code{active.assay}}{Name of the active, or default, assay; settable using \code{\link{DefaultAssay}}}

\item{\code{active.ident}}{The active cluster identity for this Seurat object; settable using \code{\link{Idents}}}

\item{\code{graphs}}{A list of \code{\link{Graph-class}} objects}

\item{\code{neighbors}}{Unused at this time}

\item{\code{reductions}}{A list of dimmensional reduction objects for this object}

\item{\code{images}}{A list of spatial image objects}

\item{\code{project.name}}{Name of the project}

\item{\code{misc}}{A list of miscellaneous information}

\item{\code{version}}{Version of Seurat this object was built under}

\item{\code{commands}}{A list of logged commands run on this \code{Seurat} object}

\item{\code{tools}}{A list of miscellaneous data generated by other tools, should be filled by developers only using \code{\link{Tool}<-}}
}}

