% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{seurat_batch_correct}
\alias{seurat_batch_correct}
\title{Batch Correct Multiple Seurat Objects}
\usage{
seurat_batch_correct(seu_list, method = "cca", ...)
}
\arguments{
\item{...}{}
}
\description{
Batch Correct Multiple Seurat Objects
}
