% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{seurat_cluster}
\alias{seurat_cluster}
\title{Run Louvain Clustering at Multiple Resolutions}
\usage{
seurat_cluster(seu, resolution = 0.6, custom_clust = NULL,
  algorithm = 1)
}
\arguments{
\item{custom_clust}{}
}
\description{
Run Louvain Clustering at Multiple Resolutions
}
