% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{seurat_cluster2}
\alias{seurat_cluster2}
\title{Run Louvain Clustering at Multiple Resolutions}
\usage{
seurat_cluster2(seu, resolution = 0.6, custom_clust = NULL)
}
\arguments{
\item{custom_clust}{}
}
\description{
Run Louvain Clustering at Multiple Resolutions
}
