% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_reference_integrate.R
\name{seurat_find_markers}
\alias{seurat_find_markers}
\title{Find Cell Type Markers in a Seurat Object}
\usage{
seurat_find_markers(seu, num_features)
}
\arguments{
\item{seu}{}
}
\description{
Find Cell Type Markers in a Seurat Object
}
