% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{seurat_integrate}
\alias{seurat_integrate}
\title{Batch Correct Multiple Seurat Objects}
\usage{
seurat_integrate(seu_list, method = "cca", organism = "human", ...)
}
\arguments{
\item{seu_list}{List of two or more seurat objects}

\item{method}{Default "cca"}

\item{...}{}
}
\description{
Batch Correct Multiple Seurat Objects
}
