% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{seurat_integration_pipeline}
\alias{seurat_integration_pipeline}
\title{Run Seurat Integration}
\usage{
seurat_integration_pipeline(seus, res_low = 0.2, res_hi = 2,
  suffix = "", ...)
}
\arguments{
\item{suffix}{}
}
\description{
run batch correction, followed by:
\enumerate{
\item stashing of batches in metadata 'batch'
\item clustering with resolution 0.2 to 2.0 in increments of 0.2
\item saving to <proj_dir>/output/sce/<feature>\emph{seu}<suffix>.rds
}
}
