% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{seurat_preprocess}
\alias{seurat_preprocess}
\title{Preprocess Seurat Object}
\usage{
seurat_preprocess(
  assay,
  scale = TRUE,
  normalize = TRUE,
  features = NULL,
  legacy_settings = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Preprocess Seurat Object
}
\examples{

panc8[["gene"]] <- seurat_preprocess(panc8[["gene"]])
}
