% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_reference_integrate.R
\name{seurat_reference_integrate}
\alias{seurat_reference_integrate}
\title{Predict One Seurat Object on another merged set}
\usage{
seurat_reference_integrate(ref_seu_list, query_seu)
}
\arguments{
\item{query_seu}{}
}
\description{
Predict One Seurat Object on another merged set
}
