% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_example_dataset}
\alias{small_example_dataset}
\title{Small example SingleCellExperiment}
\format{
An SCE with 200 cells and 1000 genes
}
\source{
scuttle::mockSCE
}
\usage{
small_example_dataset
}
\description{
created with scuttle::mockSCE
}
\keyword{datasets}
