% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{stash_batches}
\alias{stash_batches}
\title{Stash Batches in Integrated Seurat Object}
\usage{
stash_batches(corrected_seus)
}
\arguments{
\item{corrected_seus}{}
}
\description{
Stash Batches in Integrated Seurat Object
}
