% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{subset_by_meta}
\alias{subset_by_meta}
\title{Subset by new metadata}
\usage{
subset_by_meta(meta_path, object)
}
\arguments{
\item{meta_path}{Path to new metadata}

\item{object}{A object}
}
\value{
a SingleCellExperiment object
}
\description{
Subset the object using new metadata
}
