% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap_counts_from_feature}
\alias{swap_counts_from_feature}
\title{Swap counts from Feature}
\usage{
swap_counts_from_feature(cds, featureType)
}
\arguments{
\item{featureType}{}
}
\description{
Swap counts from Feature
}
