% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{tableSelected}
\alias{tableSelected}
\title{Create Table of Selected Cells}
\usage{
tableSelected(input, output, session, seu)
}
\arguments{
\item{seu}{}
}
\description{
Create Table of Selected Cells
}
