% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{threshold_monocle_genes}
\alias{threshold_monocle_genes}
\title{Threshold monocle genes}
\usage{
threshold_monocle_genes(object, cds, min_expression = 0.05)
}
\arguments{
\item{min_expression}{}
}
\description{
Threshold monocle genes
}
