% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_sce}
\alias{tiny_sce}
\title{Tiny example SingleCellExperiment}
\format{
An SCE with only expression of NRL gene and NRL transripts
}
\source{
chevreuldata::human_gene_transcript_sce()
}
\usage{
tiny_sce
}
\description{
subset to only NRL from chevreuldata::human_gene_transcript_sce()
}
\keyword{datasets}
