% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{transcripts_to_genes}
\alias{transcripts_to_genes}
\title{Ensembl Transcript Ids to Gene Symbols}
\usage{
transcripts_to_genes(transcripts, organism = "human")
}
\arguments{
\item{transcripts}{}
}
\value{

}
\description{
convert ensembl transcript ids to hgnc gene symbols
}
