% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{unite_metadata,Seurat-method}
\alias{unite_metadata,Seurat-method}
\title{Unite metadata}
\usage{
\S4method{unite_metadata}{Seurat}(object, group_bys)
}
\arguments{
\item{object}{A object}

\item{group_bys}{A feature or variable to combine}
}
\value{
an object with Idents formed from concatenation of group_bys
}
\description{
Unite metadata
}
