% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{unite_metadata}
\alias{unite_metadata}
\title{Unite metadata}
\usage{
unite_metadata(seu, metavars)
}
\arguments{
\item{seu}{A seurat object}

\item{metavars}{A feature or variable to combine}
}
\description{
Unite metadata
}
