% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_chevreul_object}
\alias{update_chevreul_object}
\title{Update a chevreul Object}
\usage{
update_chevreul_object(
  object_path,
  feature,
  resolution = seq(0.2, 2, by = 0.2),
  return_object = TRUE,
  ...
)
}
\arguments{
\item{object_path}{Path to a object}

\item{...}{}
}
\description{
Update a chevreul Object
}
