% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seurat_from_child_metadata.R
\name{update_object_meta}
\alias{update_object_meta}
\title{Update Seurat Metadata from Project}
\usage{
update_object_meta(object, proj_dir, numcols)
}
\arguments{
\item{object}{A object}

\item{proj_dir}{Project directory}

\item{numcols}{number of columns}
}
\value{
updated object metadata
}
\description{
Given a project metadata file located in <proj_dir>/data/<meta_file>, update an existing object with the project data
}
