% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seurat_from_child_metadata.R
\name{update_seu_meta}
\alias{update_seu_meta}
\title{Update Seurat Metadata}
\usage{
update_seu_meta(seu_meta, merged_new_meta, numcols)
}
\arguments{
\item{numcols}{}
}
\description{
Update Seurat Metadata
}
