% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_velocity.R
\name{velocyto_seu}
\alias{velocyto_seu}
\title{run velocyto on a seurat object}
\usage{
velocyto_seu(seu, ldat, fit.quantile = 0.05)
}
\arguments{
\item{ldat}{}
}
\value{

}
\description{
run velocyto on a seurat object
}
