% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocyto.R
\name{velocyto_seurat_from_loom}
\alias{velocyto_seurat_from_loom}
\title{Run RNA Velocity starting with only a loom File}
\usage{
velocyto_seurat_from_loom(loom_path)
}
\arguments{
\item{loom_path}{}
}
\value{

}
\description{
Run RNA Velocity starting with only a loom File
}
