% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_app.R
\name{viewSeurat}
\alias{viewSeurat}
\title{Create Seurat App}
\usage{
viewSeurat(
  seurat_object,
  feature = "gene",
  organism_type = "human",
  futureMb = 13000
)
}
\arguments{
\item{seurat_object}{a seurat object}

\item{feature}{"gene" or "transcript"}

\item{organism_type}{"human" or "mouse"}

\item{futureMb}{amount of Mb allocated to future package}
}
\value{

}
\description{
Create Seurat App
}
