% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{add_census_slot}
\alias{add_census_slot}
\title{add census assay to a object}
\usage{
add_census_slot(object, assay = "gene", slot = "counts")
}
\arguments{
\item{slot}{}
}
\description{
add census assay to a object
}
