% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{add_percent_mito}
\alias{add_percent_mito}
\title{Annotate Low Read Count Category}
\usage{
add_percent_mito(seu, organism = "human", seurat_assay = "gene")
}
\arguments{
\item{seu}{A seurat object}

\item{organism}{mouse}

\item{seurat_assay}{gene}
}
\description{
Add a Read Count Categorical Variable to Seurat Object (based on nCount_RNA)
}
\examples{
add_percent_mito(panc8)
add_percent_mito(baron2016singlecell, organism = "mouse")
}
