% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{add_percent_mito}
\alias{add_percent_mito}
\title{Annotate Low Read Count Category}
\usage{
add_percent_mito(seu, feature, organism = "human")
}
\arguments{
\item{seu}{A seurat object}
}
\value{

}
\description{
Add a Read Count Categorical Variable to Seurat Object (based on nCount_RNA)
}
