% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{add_read_count_col}
\alias{add_read_count_col}
\title{Add a 'read_count' column to a seurat object}
\usage{
add_read_count_col(seu, threshold)
}
\arguments{
\item{threshold}{}
}
\description{
Add a read count column factor to indicate 'low read count' cells below a given threshold
}
