% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{allTranscripts}
\alias{allTranscripts}
\title{Plot All Transcripts Server}
\usage{
allTranscripts(input, output, session, seu, featureType, organism_type)
}
\arguments{
\item{featureType}{}
}
\description{
Plot All Transcripts Server
}
