% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{allTranscripts}
\alias{allTranscripts}
\title{Plot All Transcripts}
\usage{
allTranscripts(input, output, session, seu, feature_type)
}
\arguments{
\item{feature_type}{}
}
\description{
Plot All Transcripts
}
