% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{annotate_cell_cycle}
\alias{annotate_cell_cycle}
\title{Annotate Cell Cycle}
\usage{
annotate_cell_cycle(object, organism = "human", ...)
}
\arguments{
\item{object}{A object}

\item{organism}{organism. Default "human"}
}
\value{
a single cell object
}
\description{
Annotate Cell Cycle for Gene and Transcript Seurat Objects
}
