% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{annotate_cell_cycle}
\alias{annotate_cell_cycle}
\title{Annotate Cell Cycle}
\usage{
annotate_cell_cycle(seu, organism = "human", ...)
}
\arguments{
\item{organism}{}
}
\description{
Annotate Cell Cycle for Gene and Transcript Seurat Objects
}
\examples{
annotate_cell_cycle(panc8, organism = "human")
annotate_cell_cycle(baron2016singlecell, organism = "mouse")
}
