% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{annotate_cell_cycle}
\alias{annotate_cell_cycle}
\title{Annotate Cell Cycle}
\usage{
annotate_cell_cycle(seu, s.features, g2m.features, ...)
}
\arguments{
\item{seu_list}{A list of seurat objects}
}
\description{
Annotate Cell Cycle for Gene and Transcript Seurat Objects
}
