% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{annotate_cell_cycle}
\alias{annotate_cell_cycle}
\title{Annotate Cell Cycle}
\usage{
annotate_cell_cycle(object)
}
\arguments{
\item{object}{A single cell object}
}
\value{
a single cell object
}
\description{
Annotate Cell Cycle for Gene and Transcript SingleCellExperiment Objects
}
\examples{
human_gene_phase<-annotate_cell_cycle(human_gene_transcript_sce)
human_gene_phase$phases

}
