% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_to_project_db}
\alias{append_to_project_db}
\title{Update a database of chevreul projects}
\usage{
append_to_project_db(
  new_project_path,
  cache_location = "~/.cache/chevreul",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{new_project_path}{new project path}

\item{cache_location}{Path to cache "~/.cache/chevreul"}

\item{sqlite_db}{sqlite db}

\item{verbose}{print messages}
}
\value{
a sqlite database with SingleCellExperiment objects
}
\description{
Append projects to database
}
