% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{build_bigwig_db}
\alias{build_bigwig_db}
\title{Create a database of bigwigfiles}
\usage{
build_bigwig_db(bam_files, bigwig_db = "~/.cache/seuratTools/bw-files.db")
}
\arguments{
\item{bigwig_db}{}
}
\description{
create a sqlite database of bigwig files matching cell ids in seurat objects
}
