% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{celltype_markers}
\alias{celltype_markers}
\title{Retinal Cell Type Marker Genes}
\format{
A data frame with 99 rows and 3 variables:
\describe{
\item{\code{gene}}{character COLUMN_DESCRIPTION}
\item{\code{cell_type}}{character COLUMN_DESCRIPTION}
\item{\code{species}}{character COLUMN_DESCRIPTION}
}
}
\source{
Cui et al. (2020). Transcriptomic Analysis of the Developmental Similarities and Differences Between the Native Retina and Retinal Organoids. Invest. Ophthalmol. Vis. Sci. 61, 6–6.
}
\usage{
celltype_markers
}
\description{
Retinal Cell Type Marker Genes
}
\keyword{datasets}
