% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevreul-package.R
\docType{package}
\name{chevreul-package}
\alias{chevreul}
\alias{chevreul-package}
\title{chevreul: Tools for managing SingleCellExperiment objects as projects}
\description{
Tools for managing SingleCellExperiment objects as projects. Includes functions for analysis and visualization of single-cell data. Also included is a shiny app for visualization of pre-processed scRNA data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whtns/chevreul}
  \item \url{https://whtns.github.io/chevreul/}
  \item Report bugs at \url{https://github.com/cobriniklab/chevreul/issues}
}

}
\author{
\strong{Maintainer}: Kevin Stachelek \email{kevin.stachelek@gmail.com} (\href{https://orcid.org/0000-0003-2085-695X}{ORCID})

Authors:
\itemize{
  \item Bhavana Bhat \email{bhavanabhat29@gmail.com}
}

}
\keyword{internal}
