% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevreul_app.R
\name{chevreulApp}
\alias{chevreulApp}
\title{Create a shiny app for a project on disk}
\usage{
chevreulApp(
  preset_project,
  appTitle = "chevreul",
  organism_type = "human",
  db_path = "~/.cache/chevreul/single-cell-projects.db",
  futureMb = 13000,
  integrated_proj_dir = "/dataVolume/storage/single_cell_projects/integrated_projects/"
)
}
\arguments{
\item{preset_project}{A preloaded project to start the app with}

\item{appTitle}{A title of the App}

\item{organism_type}{human or mouse}

\item{db_path}{sqlite database with list of saved SingleCellExperiment objects}

\item{futureMb}{amount of Mb allocated to future package}

\item{integrated_proj_dir}{Path to integrated project directory}
}
\value{
a shiny app
}
\description{
Create a shiny app for a project on disk
}
\examples{
\donttest{
chevreulApp()
}

}
