% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{chevreulApp}
\alias{chevreulApp}
\title{Create a shiny app for a project on disk}
\usage{
chevreulApp(
  preset_project,
  appTitle = "chevreul",
  organism_type = "human",
  db_path = "~/.cache/chevreul/single-cell-projects.db",
  futureMb = 13000
)
}
\arguments{
\item{preset_project}{preset project}

\item{appTitle}{A title of the App}

\item{organism_type}{human or mouse}

\item{db_path}{sqlite database with list of saved single cell objects}

\item{futureMb}{megabytes for use with future package}
}
\value{
a shiny app
}
\description{
Create a shiny app for a project on disk
}
