% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chevreul_app.R
\name{chevreulApp}
\alias{chevreulApp}
\title{Create a shiny app for a project on disk}
\usage{
chevreulApp(
  preset_project,
  appTitle = "chevreul",
  organism_type = "human",
  futureMb = 13000,
  db_name = "single-cell-projects.db"
)
}
\arguments{
\item{preset_project}{A preloaded project to start the app with}

\item{appTitle}{A title of the App}

\item{organism_type}{human or mouse}

\item{futureMb}{amount of Mb allocated to future package}

\item{db_name}{sqlite database with list of saved
SingleCellExperiment objects}
}
\value{
a shiny app
}
\description{
Create a shiny app for a project on disk
}
