% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{clustering_workflow}
\alias{clustering_workflow}
\title{Clustering Workflow}
\usage{
clustering_workflow(
  object,
  excluded_cells,
  resolution = seq(0.2, 2, by = 0.2),
  organism = "human",
  experiment_name = "default_experiment",
  ...
)
}
\arguments{
\item{excluded_cells}{named list of cells to exclude}

\item{resolution}{resolution(s) to use for clustering cells}

\item{organism}{Organism}

\item{...}{}

\item{feature_objects}{list of objects named according to feature of interest ("gene" or "transcript")}
}
\description{
Cluster and Reduce Dimensions of a object
}
