% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{clustering_workflow}
\alias{clustering_workflow}
\title{Clustering Workflow}
\usage{
clustering_workflow(
  feature_seus = NULL,
  excluded_cells,
  cell_cycle = FALSE,
  resolution = seq(0.2, 2, by = 0.2),
  organism = "human",
  experiment_name = "batch",
  ...
)
}
\arguments{
\item{feature_seus}{list of seurat objects named according to feature of interest ("gene" or "transcript")}

\item{excluded_cells}{named list of cells to exclude}

\item{cell_cycle}{whether to score and regress cell cycle related features}

\item{resolution}{resolution(s) to use for clustering cells}

\item{...}{}
}
\value{

}
\description{
Integrate multiple seurat objects and save to file
}
