% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{combine_looms}
\alias{combine_looms}
\title{Combine Loom Files}
\usage{
combine_looms(projectPaths, newProjectPath)
}
\arguments{
\item{projectPaths}{project paths}

\item{newProjectPath}{new project path}
}
\value{
a combined loom file
}
\description{
Combine Loom Files
}
