% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_symbols_by_species}
\alias{convert_symbols_by_species}
\title{Convert gene symbols between mouse and human}
\usage{
convert_symbols_by_species(src_genes, src_species)
}
\arguments{
\item{src_species}{}
}
\value{

}
\description{
Convert gene symbols between mouse and human
}
\examples{

convert_symbols_by_species("RXRG", "human")

convert_symbols_by_species("Rxrg", "mouse")

}
