% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{convert_symbols_by_species}
\alias{convert_symbols_by_species}
\title{Convert gene symbols between mouse and human}
\usage{
convert_symbols_by_species(src_genes, src_species)
}
\arguments{
\item{src_genes}{Source gene symbol to be converted}

\item{src_species}{Source species}
}
\value{
a SingleCellExperiment object
}
\description{
Convert gene symbols between mouse and human
}
