% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{convert_to_h5ad}
\alias{convert_to_h5ad}
\title{convert a seurat object to an on-disk anndata object}
\usage{
convert_to_h5ad(seu, file_path)
}
\arguments{
\item{file_path}{}
}
\value{

}
\description{
convert a seurat object to an on-disk anndata object
}
