% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{convert_to_h5ad}
\alias{convert_to_h5ad}
\title{convert a object to an on-disk anndata object}
\usage{
convert_to_h5ad(object, file_path)
}
\arguments{
\item{object}{A object}

\item{file_path}{Path to file}
}
\description{
convert a object to an on-disk anndata object
}
\examples{

convert_to_h5ad(human_gene_transcript_object, "inst/extdata/object.rds")

}
