% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_v3_to_v5}
\alias{convert_v3_to_v5}
\title{Convert seurat to seurat V5 format}
\usage{
convert_v3_to_v5(seurat_v3)
}
\arguments{
\item{seurat_v3}{a version 3 seurat}
}
\description{
Convert seurat from v3 to v5 format
}
\examples{
convert_v3_to_v5(human_gene_transcript_seurat)
}
