% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_project_db}
\alias{create_project_db}
\title{Create a database of chevreul projects}
\usage{
create_project_db(
  cache_location = "~/.cache/chevreul",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Create a database of chevreul projects
}
