% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_project_db}
\alias{create_project_db}
\title{Create a database of seuratTools projects}
\usage{
create_project_db(
  destdir = "~/.cache/seuratTools",
  destfile = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Create a database of seuratTools projects
}
