% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_project_db}
\alias{create_project_db}
\title{Create a database of chevreul projects}
\usage{
create_project_db(
  cache_location = "~/.cache/chevreul",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{cache_location}{Path to cache "~/.cache/chevreul"}

\item{sqlite_db}{Database to be created}

\item{verbose}{print messages}
}
\value{
a sqlite database with SingleCellExperiment objects
}
\description{
Create a database containing chevreul projects
}
