% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{cross_species_integrate}
\alias{cross_species_integrate}
\title{Integrate Seurat Objects from Mouse to Human}
\usage{
cross_species_integrate(
  mouse_seu_list,
  human_seu_list,
  excluded_cells = NULL,
  ...
)
}
\arguments{
\item{human_seu_list}{}
}
\value{

}
\description{
Integrate Seurat Objects from Mouse to Human
}
\examples{



}
