% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{enframe_markers}
\alias{enframe_markers}
\title{Enframe Markers}
\usage{
enframe_markers(marker_table)
}
\arguments{
\item{marker_table}{a table of marker genes}
}
\value{
a table of marker genes
}
\description{
Enframe Markers
}
