% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{filter_merged_object}
\alias{filter_merged_object}
\title{Filter a Single Seurat Object}
\usage{
filter_merged_object(object, filter_var, filter_val, .drop = .drop)
}
\arguments{
\item{object}{A singlecell objects}

\item{.drop}{}
}
\value{
a single cell object
}
\description{
Filter a Single Seurat Object
}
