% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{filter_merged_objects}
\alias{filter_merged_objects}
\title{Filter a List of Seurat Objects}
\usage{
filter_merged_objects(objects, filter_var, filter_val, .drop = F)
}
\arguments{
\item{.drop}{}
}
\description{
Filter Seurat Objects by custom variable and reset assay to uncorrected "gene"
}
