% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{find_all_markers}
\alias{find_all_markers}
\title{Find All Markers}
\usage{
find_all_markers(object, group_by = NULL, experiment = "gene", ...)
}
\arguments{
\item{object}{An object.}

\item{group_by}{A metadata variable to group by.}

\item{experiment}{Assay to use, Default "gene".}

\item{...}{extra args passed to stash_marker_features}
}
\value{
a SingleCellExperiment object containing marker genes
}
\description{
Find all markers at a range of resolutions
}
