% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{find_all_markers}
\alias{find_all_markers}
\title{Find All Markers}
\usage{
find_all_markers(seu, metavar = NULL, seurat_assay = "gene", ...)
}
\arguments{
\item{seu}{A seurat object.}

\item{metavar}{A metadata variable to group by.}

\item{seurat_assay}{Assay to use, Default "gene".}

\item{...}{}
}
\description{
Find all markers at a range of resolutions
}
\examples{
markers_stashed_seu <- find_all_markers(panc8)
marker_genes <- Misc(markers_stashed_seu, "markers")
str(marker_genes)
}
