% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{find_all_markers}
\alias{find_all_markers}
\title{Find All Markers}
\usage{
find_all_markers(object, metavar = NULL, object_assay = "gene", ...)
}
\arguments{
\item{object}{An object.}

\item{metavar}{A metadata variable to group by.}

\item{object_assay}{Assay to use, Default "gene".}

\item{...}{}
}
\description{
Find all markers at a range of resolutions
}
\examples{
markers_stashed_object <- find_all_markers(panc8)
marker_genes <- Misc(markers_stashed_object, "markers")
str(marker_genes)
}
