% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{gene_test_plot}
\alias{gene_test_plot}
\title{Plot Gene expression over pseudotime on reference and query trajectories from cellAlign}
\usage{
gene_test_plot(
  expGlobalRef,
  expGlobalQuery,
  trajRef,
  trajQuery,
  winSz = 0.1,
  numPts = 200
)
}
\arguments{
\item{expGlobalRef}{reference experiment}

\item{expGlobalQuery}{query experiment}

\item{trajRef}{reference trajectory}

\item{trajQuery}{query trajectory}

\item{winSz}{window size}

\item{numPts}{number of points}
}
\value{
a plot
}
\description{
Plot Gene expression over pseudotime on reference and query trajectories from cellAlign
}
