% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{genes_to_transcripts}
\alias{genes_to_transcripts}
\title{Ensembl Genes to Transcripts}
\usage{
genes_to_transcripts(genes)
}
\arguments{
\item{genes}{}
}
\description{
convert genes to transcripts
}
