% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_cell_metadata}
\alias{get_cell_metadata}
\title{Pull the metadata from a given object}
\usage{
get_cell_metadata(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
dataframe containing object metadata
}
\description{
Pull the metadata from a given object
}
\examples{
get_cell_metadata(human_gene_transcript_sce)
}
