% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_feature_types}
\alias{get_feature_types}
\title{Get Feature Types}
\usage{
get_feature_types(object)
}
\arguments{
\item{object}{}
}
\description{
Get Feature Types
}
\examples{
get_feature_types(human_gene_transcript_sce)
}
