% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_features}
\alias{get_features}
\title{Get feature names}
\usage{
get_features(object, experiment = "gene")
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{experiment}{"gene" or "transcript"}
}
\value{
variable features from a SingleCellExperiment object
}
\description{
Get feature names
}
\examples{

data(small_example_dataset)
get_features(small_example_dataset)
}
