% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_object_metadata}
\alias{get_object_metadata}
\title{Get object metadata}
\usage{
get_object_metadata(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
variable features from a single cell object
}
\description{
Get object metadata
}
\examples{
get_object_metadata(human_gene_transcript_sce)
}
