% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transcripts_from_object}
\alias{get_transcripts_from_object}
\title{Get Transcripts in object}
\usage{
get_transcripts_from_object(object, gene, organism = "human")
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{gene}{Gene of interest}

\item{organism}{Organism}
}
\value{
transcripts constituting a gene of interest in a single cell object
}
\description{
Get transcript ids in objects for one or more gene of interest
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()

NRL_transcripts <- get_transcripts_from_object(chevreul_sce, "NRL")

}
