% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transcripts_from_seu}
\alias{get_transcripts_from_seu}
\title{Get Transcripts in Seurat Object}
\usage{
get_transcripts_from_seu(seu, gene, organism = "human")
}
\arguments{
\item{organism}{}
}
\description{
Get Transcripts in Seurat Object
}
\examples{
RXRG_transcripts <- get_transcripts_from_seu(human_gene_transcript_seu, "RXRG")
}
