% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classes.R
\name{get_variable_features}
\alias{get_variable_features}
\title{Pull object metadata}
\usage{
get_variable_features(object, ...)
}
\arguments{
\item{object}{}
}
\value{
variable features from a single cell object
}
\description{
Pull object metadata
}
