% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{get_variable_features}
\alias{get_variable_features}
\title{Get variable features}
\usage{
get_variable_features(object, experiment = "gene")
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{experiment}{"gene" or "transcript"}
}
\value{
variable features from a SingleCellExperiment object
}
\description{
Get variable features
}
\examples{

data(small_example_dataset)
get_variable_features(small_example_dataset)
}
