% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{global_cellalign}
\alias{global_cellalign}
\title{Global Alignment of two trajectories with cellAlign}
\usage{
global_cellalign(exp_ref, exp_query, traj_ref, traj_query, numPts = 200)
}
\arguments{
\item{exp_ref}{reference experiment}

\item{exp_query}{query experiment}

\item{traj_ref}{reference trajectory}

\item{traj_query}{query trajectory}
}
\value{
a list containing an alignment and an alignment plot
}
\description{
Global Alignment of two trajectories with cellAlign
}
