% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{harmony_integrate}
\alias{harmony_integrate}
\title{Integrate small datasets with harmony}
\usage{
harmony_integrate(object_list)
}
\arguments{
\item{object_list}{List of two or more singlecell objectss to integrate}
}
\value{
an integrated single cell object
}
\description{
Integrate small datasets with harmony
}
