% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{integration_workflow}
\alias{integration_workflow}
\title{Integration Workflow}
\usage{
integration_workflow(
  batches,
  excluded_cells = NULL,
  resolution = seq(0.2, 2, by = 0.2),
  experiment_name = "default_experiment",
  organism = "human",
  ...
)
}
\arguments{
\item{batches}{objects for all batches provided as a list. If named, the resulting integrated object will be identified with corresponding values in 'batch' metadata}

\item{excluded_cells}{named list of cells to exclude}

\item{resolution}{value(s) to control the clustering resolution via \code{SingleCellExperiment::FindMarkers}}

\item{experiment_name}{arbitrary name to identify experiment}

\item{organism}{either "human" or "mouse"}

\item{...}{extra args passed to object_integration_pipeline}
}
\value{
an integrated single cell object
}
\description{
Integrate multiple objects and save to file
}
\examples{
batches <- splitByCol(human_gene_transcript_sce, "batch")
integration_workflow(batches)
}
