% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{integration_workflow}
\alias{integration_workflow}
\title{Integration Workflow}
\usage{
integration_workflow(child_proj_dirs, excluded_cells, ...)
}
\arguments{
\item{child_proj_dirs}{child projects to be integrated}

\item{excluded_cells}{named list of cells to exclude}

\item{...}{}
}
\description{
Integrate multiple seurat objects and save to file
}
