% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_reference_integrate.R
\name{label_transfer}
\alias{label_transfer}
\title{Transfer Labels Between Seurat Objects}
\usage{
label_transfer(
  ref_object,
  query_object,
  ref_name = NULL,
  query_name = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Transfer Labels Between Seurat Objects
}
