% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{list_plot_types}
\alias{list_plot_types}
\title{Collate list of variables to be plotted}
\usage{
list_plot_types(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
plot_types a list of category_vars or continuous_vars
}
\description{
Collate list of variables to be plotted
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
list_plot_types(chevreul_sce)
}
