% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{load_bigwigs}
\alias{load_bigwigs}
\title{Load Bigwigs}
\usage{
load_bigwigs(seu, bigwig_db = "~/.cache/seuratTools/bw-files.db")
}
\arguments{
\item{bigwig_db}{}
}
\value{

}
\description{
Load a tibble of bigwig file paths by cell id
}
