% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{load_bigwigs}
\alias{load_bigwigs}
\title{Load Bigwigs}
\usage{
load_bigwigs(object, bigwig_db = "~/.cache/chevreul/bw-files.db")
}
\arguments{
\item{object}{A object}

\item{bigwig_db}{Sqlite database of bigwig files}
}
\value{
a vector of bigwigs file paths
}
\description{
Load a tibble of bigwig file paths by cell id
}
