% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{load_counts_by_tximport}
\alias{load_counts_by_tximport}
\title{Run \href{http://bioconductor.org/packages/release/bioc/html/tximport.html}{tximport} on a set of cells}
\usage{
load_counts_by_tximport(
  proj_dir,
  type = "salmon",
  countsFromAbundance = "scaledTPM",
  edb = EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86
)
}
\arguments{
\item{proj_dir}{project directory}

\item{type}{stringtie or salmon}

\item{countsFromAbundance}{argument provided to tximport}

\item{edb}{}
}
\value{

}
\description{
cells can be quantified using:
\itemize{
\item stringtie
\item salmon
}
}
