% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{load_counts_from_stringtie}
\alias{load_counts_from_stringtie}
\title{Load Counts from Stringtie Files}
\usage{
load_counts_from_stringtie(proj_dir, txOut)
}
\arguments{
\item{txOut}{}
}
\description{
Load Counts from Stringtie Files
}
