% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellalign.R
\name{local_cellalign}
\alias{local_cellalign}
\title{Align two trajectories via cellalign}
\usage{
local_cellalign(
  exp_ref,
  exp_query,
  traj_ref,
  traj_query,
  Thresh = 0.2,
  winSz = 0.1,
  numPts = 200
)
}
\arguments{
\item{numPts}{}
}
\value{

}
\description{
Align two trajectories via cellalign
}
