% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_bigwig_db}
\alias{make_bigwig_db}
\title{Make Bigwig Database}
\usage{
make_bigwig_db(destdir = "~/.cache/seuratTools/", destfile = "bw-files.db")
}
\arguments{
\item{destfile}{}
}
\value{

}
\description{
Make Bigwig Database
}
