% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_bigwig_db}
\alias{make_bigwig_db}
\title{Make Bigwig Database}
\usage{
make_bigwig_db(
  new_project = NULL,
  cache_location = "~/.cache/chevreul/",
  sqlite_db = "bw-files.db"
)
}
\arguments{
\item{new_project}{Project directory}

\item{cache_location}{Path to cache "~/.cache/chevreul"}

\item{sqlite_db}{sqlite db containing bw files}
}
\value{
a sqlite database of bigwig files for cells
in a SingleCellExperiment object
}
\description{
Make Bigwig Database
}
