% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{merge_small_objects}
\alias{merge_small_objects}
\title{Merge Small Seurat Objects}
\usage{
merge_small_objects(object_list, k.filter = 50)
}
\arguments{
\item{object_list}{List of two or more singlecell objects}

\item{k.filter}{minimum cell number for integration}
}
\value{
a single cell object
}
\description{
Merge Small Seurat Objects
}
