% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{merge_small_objects}
\alias{merge_small_objects}
\title{Merge Small SingleCellExperiment Objects}
\usage{
merge_small_objects(..., k.filter = 50)
}
\arguments{
\item{...}{two or more singlecell objects}

\item{k.filter}{minimum cell number for integration}
}
\value{
a SingleCellExperiment object
}
\description{
Merge Small SingleCellExperiment Objects
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
merge_small_objects(
    "small_batch1" = chevreul_sce[, seq(1, 40)],
    "small_batch2" = chevreul_sce[, seq(41, 80)],
    "large_batch" = chevreul_sce[, seq(81, 300)]
)

}
