% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metadata_from_batch}
\alias{metadata_from_batch}
\title{Retrieve Metadata from Batch}
\usage{
metadata_from_batch(
  batch,
  projects_dir = "/dataVolume/storage/single_cell_projects",
  db_path = "single-cell-projects.db"
)
}
\arguments{
\item{db_path}{}
}
\value{

}
\description{
Retrieve Metadata from Batch
}
