% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metadata_from_batch}
\alias{metadata_from_batch}
\title{Retrieve Metadata from Batch}
\usage{
metadata_from_batch(
  batch,
  projects_dir = "/dataVolume/storage/single_cell_projects",
  db_path = "single-cell-projects.db"
)
}
\arguments{
\item{batch}{batch}

\item{projects_dir}{path to project dir}

\item{db_path}{path to .db file}
}
\value{
a tibble with cell level metadata from a SingleCellExperiment object
}
\description{
Retrieve Metadata from Batch
}
