% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_chevreul_app.R
\name{minimalChevreulApp}
\alias{minimalChevreulApp}
\title{Create a minimal chevreul app using SingleCellExperiment input}
\usage{
minimalChevreulApp(
  single_cell_object = chevreul_sce,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000,
  db_name = "single-cell-projects.db"
)
}
\arguments{
\item{single_cell_object}{a singlecell object}

\item{appTitle}{a title for the app}

\item{organism_type}{human or mouse}

\item{futureMb}{the megabytes available for the future package}

\item{db_name}{a database of bigwig files}
}
\value{
a minimal chevreul app
}
\description{
Create a minimal chevreul app using SingleCellExperiment input
}
\examples{
\donttest{


chevreul_sce <- small_example_dataset
minimalChevreulApp(chevreul_sce)
}

}
