% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_chevreul_app.R
\name{minimalChevreulApp}
\alias{minimalChevreulApp}
\title{Create a minimal chevreul app using SingleCellExperiment input}
\usage{
minimalChevreulApp(
  single_cell_object = chevreul_sce,
  loom_path = NULL,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000
)
}
\arguments{
\item{single_cell_object}{a singlecell object}

\item{loom_path}{path to a loom file}

\item{appTitle}{a title for the app}

\item{organism_type}{human or mouse}

\item{futureMb}{the megabytes available for the future package}

\item{bigwig_db}{a database of bigwig files}
}
\value{
a minimal chevreul app
}
\description{
Create a minimal chevreul app using SingleCellExperiment input
}
\examples{
\donttest{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
minimalChevreulApp(chevreul_sce)
}

}
