% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_sce_app.R
\name{minimalSceApp}
\alias{minimalSceApp}
\title{Create a minimal chevreul app using SingleCellExperiment input}
\usage{
minimalSceApp(
  input_object = panc8,
  loom_path = NULL,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000,
  bigwig_db = "~/.cache/chevreul/bw-files.db"
)
}
\arguments{
\item{loom_path}{path to a loom file}

\item{appTitle}{a title for the app}

\item{organism_type}{human or mouse}

\item{futureMb}{the megabytes available for the future package}

\item{object}{a singlecell object}
}
\value{
a minimal chevreul app
}
\description{
Create a minimal chevreul app using SingleCellExperiment input
}
\examples{
\dontrun{
minimalSceApp(panc8)
}

}
