% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_app.R
\name{minimalSeuratApp}
\alias{minimalSeuratApp}
\title{Create a minimal seurat app}
\usage{
minimalSeuratApp(
  object = seurat_pancreas_reduced,
  loom_path = NULL,
  appTitle = NULL,
  feature_types = "gene",
  organism_type = "human",
  futureMb = 13000
)
}
\arguments{
\item{futureMb}{}
}
\value{

}
\description{
Create a minimal seurat app
}
