% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_seu_app.R
\name{minimalSeuratApp}
\alias{minimalSeuratApp}
\title{Create a minimal object app}
\usage{
minimalSeuratApp(
  single_cell_object = panc8,
  loom_path = NULL,
  appTitle = NULL,
  organism_type = "human",
  futureMb = 13000,
  bigwig_db = "~/.cache/chevreul/bw-files.db"
)
}
\arguments{
\item{single_cell_object}{a single cell object}

\item{loom_path}{path to a loom file}

\item{appTitle}{title for an app}

\item{organism_type}{Organism}

\item{futureMb}{megabytes available for future package}
}
\value{
a minimal chevreul app
}
\description{
Create a minimal object app
}
\examples{
\dontrun{
minimalSeuratApp(panc8)
}

}
