% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_app.R
\name{minimalSeuratApp}
\alias{minimalSeuratApp}
\title{Create Seurat App}
\usage{
minimalSeuratApp(
  seu,
  appTitle = NULL,
  feature_types = "gene",
  organism_type = "human",
  loom_path = NULL,
  futureMb = 13000
)
}
\arguments{
\item{seu}{A preloaded project to start the app with}

\item{appTitle}{A title of the App}

\item{futureMb}{amount of Mb allocated to future package}
}
\value{

}
\description{
Create Seurat App
}
