% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{monocle_module_heatmap}
\alias{monocle_module_heatmap}
\title{Title}
\usage{
monocle_module_heatmap(
  cds,
  pr_deg_ids,
  seu_resolution,
  cells = NULL,
  collapse_rows = TRUE,
  resolution = 10^seq(-6, -1),
  group.by = "batch",
  group.bar.height = 0.01,
  cluster_columns = FALSE,
  cluster_rows = TRUE,
  column_split = NULL,
  col_dendrogram = "ward.D2",
  mm_col_dend = 30,
  min_percent = 0.05,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
