% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_calcn}
\alias{object_calcn}
\title{Calculate Read Count Metrics for a object}
\usage{
object_calcn(object)
}
\arguments{
\item{object}{A single cell object}
}
\value{
a single cell object with nfeatures and ngenes stored in metadata
}
\description{
Recalculate counts/features per cell for a object
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
object_calcn(chevreul_sce)
}
