% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{object_cluster}
\alias{object_cluster}
\title{Run Louvain Clustering at Multiple Resolutions}
\usage{
object_cluster(
  object = object,
  resolution = 0.6,
  custom_clust = NULL,
  reduction = "pca",
  algorithm = 1,
  ...
)
}
\arguments{
\item{object}{A single cell objects}

\item{resolution}{Clustering resolution}

\item{custom_clust}{custom cluster}

\item{reduction}{Set dimensional reduction object}

\item{algorithm}{1}

\item{...}{extra args passed to single cell packages}
}
\value{
a single cell object with louvain clusters
}
\description{
Run Louvain Clustering at Multiple Resolutions
}
