% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{object_integrate}
\alias{object_integrate}
\title{Batch Correct Multiple Seurat Objects}
\usage{
object_integrate(object_list, method = "cca", organism = "human", ...)
}
\arguments{
\item{object_list}{List of two or more singlecell objectss}

\item{method}{Default "cca"}

\item{...}{}
}
\description{
Batch Correct Multiple Seurat Objects
}
