% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{object_integrate}
\alias{object_integrate}
\title{Batch Correct Multiple Single Cell Objects}
\usage{
object_integrate(object_list, organism = "human", ...)
}
\arguments{
\item{object_list}{List of two or more single cell objects}

\item{organism}{human or mouse}

\item{...}{extra args passed to object_reduce_dimensions}

\item{method}{Default "cca"}
}
\value{
an integrated single cell object
}
\description{
Batch Correct Multiple Single Cell Objects
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
batches <- splitByCol(chevreul_sce, "batch")
object_integrate(batches)
}
