% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{object_pipeline}
\alias{object_pipeline}
\title{Run SingleCellExperiment Pipeline}
\usage{
object_pipeline(
  object,
  assay = "gene",
  resolution = 0.6,
  reduction = "PCA",
  organism = "human",
  ...
)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{assay}{Assay of interest in SingleCellExperiment object}

\item{resolution}{Resolution for clustering cells. Default set to 0.6.}

\item{reduction}{Dimensional reduction object}

\item{organism}{Organism}

\item{...}{extra parameters passed to internal functions}
}
\value{
a processed single cell object
}
\description{
This functions allows you to Preprocess, Cluster and Reduce Dimensions for a single object.
}
\examples{

processed_object <- object_pipeline(panc8)

}
