% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{object_preprocess}
\alias{object_preprocess}
\title{Preprocess Single Cell Object}
\usage{
object_preprocess(
  assay,
  scale = TRUE,
  normalize = TRUE,
  features = NULL,
  legacy_settings = FALSE,
  ...
)
}
\arguments{
\item{assay}{Assay to use}

\item{scale}{Perform linear transformation 'Scaling'}

\item{normalize}{Perform normalization}

\item{features}{Identify highly variable features}

\item{legacy_settings}{Use legacy settings}

\item{...}{}
}
\description{
Performs standard pre-processing workflow for scRNA-seq data
}
\examples{

panc8[["gene"]] <- object_preprocess(panc8[["gene"]])

}
