% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_all_transcripts}
\alias{plot_all_transcripts}
\title{Plot All Transcripts}
\usage{
plot_all_transcripts(seu, features, embedding = "umap")
}
\arguments{
\item{transcripts}{}
}
\value{

}
\description{
Plot All Transcripts
}
\examples{

Plot all transcripts

}
