% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_all_transcripts}
\alias{plot_all_transcripts}
\title{Title}
\usage{
plot_all_transcripts(seu_transcript, seu_gene, features, embedding)
}
\arguments{
\item{transcripts}{}
}
\value{

}
\description{
Title
}
