% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_all_transcripts}
\alias{plot_all_transcripts}
\title{Plot All Transcripts}
\usage{
plot_all_transcripts(
  seu,
  features,
  embedding = "umap",
  from_gene = TRUE,
  combine = TRUE
)
}
\arguments{
\item{features}{gene or vector of transcripts}

\item{embedding}{umap}

\item{from_gene}{whether to look up transcripts for an input gene}
}
\description{
plot expression all transcripts for an input gene superimposed on an embedding
}
\examples{

processed_seu <- clustering_workflow(human_gene_transcript_seu)
transcripts_to_plot <- genes_to_transcripts("RXRG")
plot_all_transcripts(processed_seu, features = transcripts_to_plot)

}
