% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R, R/monocle3.R
\name{plot_cds}
\alias{plot_cds}
\title{Plot a Monocle Cell Data Set}
\usage{
plot_cds(cds, resolution, color_cells_by = "louvain_cluster")

plot_cds(cds, resolution, color_cells_by = "louvain_cluster")
}
\arguments{
\item{color_cells_by}{}

\item{color_cells_by}{}
}
\description{
Plot a Monocle Cell Data Set

Plot a Monocle Cell Data Set
}
