% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_feature}
\alias{plot_feature}
\title{Plot Feature}
\usage{
plot_feature(
  seu,
  embedding = c("umap", "pca", "tsne"),
  features,
  dims = c(1, 2),
  return_plotly = FALSE,
  pt.size = 1
)
}
\arguments{
\item{seu}{A Seurat object}

\item{embedding}{Dimensional reduction technique to be used}

\item{features}{Features to plot}

\item{dims}{Dimensions to plot, must be a two-length numeric vector}
}
\description{
Plots gene or transcript expression overlaid on a given embedding.
If multiple features are supplied the joint density of all features
will be plotted using \href{https://www.bioconductor.org/packages/devel/bioc/html/Nebulosa.html}{Nebulosa}
}
\examples{

# static, single feature
plot_feature(human_gene_transcript_seu, embedding = "umap", features = c("NRL"), return_plotly = FALSE)
# static, multi-feature
plot_feature(human_gene_transcript_seu, embedding = "umap", features = c("RXRG", "NRL"), return_plotly = FALSE)
# interactive, multi-feature
plotly_plot <- plot_feature(human_gene_transcript_seu, embedding = "umap", features = c("RXRG", "NRL"))
print(plotly_plot)

}
