% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{plot_feature}
\alias{plot_feature}
\title{Plot Features
This is a great function}
\usage{
plot_feature(seu, embedding, features, dims = c(1, 2),
  return_plotly = TRUE)
}
\arguments{
\item{dims}{}
}
\description{
Plot Features
This is a great function
}
