% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_feature}
\alias{plot_feature}
\title{Plot Feature}
\usage{
plot_feature(
  object,
  embedding = c("UMAP", "PCA", "TSNE"),
  features,
  dims = c(1, 2),
  return_plotly = FALSE,
  pt.size = 1
)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{embedding}{Dimensional reduction technique to be used}

\item{features}{Features to plot}

\item{dims}{Dimensions to plot, must be a two-length numeric vector}

\item{return_plotly}{return plotly object}

\item{pt.size}{size of points in ggplot}
}
\value{
an embedding colored by a feature of interest
}
\description{
Plots gene or transcript expression overlaid on a given embedding.
If multiple features are supplied the joint density of all features
will be plotted using \href{https://www.bioconductor.org/packages/devel/bioc/html/Nebulosa.html}{Nebulosa}
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
plot_feature(chevreul_sce, embedding = "UMAP", features = "NRL")

}
