% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{plot_gene_coverage_by_var}
\alias{plot_gene_coverage_by_var}
\title{Plot BigWig Coverage for Genes of Interest Colored by a Given Variable}
\usage{
plot_gene_coverage_by_var(
  genes_of_interest = "RXRG",
  cell_metadata,
  bigwig_tbl,
  var_of_interest = "batch",
  values_of_interest = NULL,
<<<<<<< HEAD
  edb = EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86,
  heights = c(2, 1),
=======
  organism = "human",
  edb = NULL,
  heights = c(3, 1),
  scale_y = "log10",
  reverse_x = FALSE,
  start = NULL,
  end = NULL,
>>>>>>> master
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot BigWig Coverage for Genes of Interest Colored by a Given Variable
}
