% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{plot_gene_coverage_by_var}
\alias{plot_gene_coverage_by_var}
\title{Plot BigWig Coverage for Genes of Interest Colored by a Given Variable}
\usage{
plot_gene_coverage_by_var(
  genes_of_interest = "RXRG",
  cell_metadata_file,
  var_of_interest,
  edb = EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86
)
}
\arguments{
\item{edb}{}
}
\value{

}
\description{
Plot BigWig Coverage for Genes of Interest Colored by a Given Variable
}
