% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggleplotr.R
\name{plot_gene_coverage_by_var}
\alias{plot_gene_coverage_by_var}
\title{Plot BigWig Coverage for Genes of Interest Colored by a Given Variable}
\usage{
plot_gene_coverage_by_var(
  genes_of_interest = "RXRG",
  cell_metadata,
  bigwig_tbl,
  var_of_interest = NULL,
  values_of_interest = NULL,
  edb = EnsDb.Hsapiens.v86::EnsDb.Hsapiens.v86,
  heights = c(2, 1),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot BigWig Coverage for Genes of Interest Colored by a Given Variable
}
