% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_markers}
\alias{plot_markers}
\title{Plot Cluster Marker Genes}
\usage{
plot_markers(
  seu,
  metavar = "batch",
  num_markers = 5,
  selected_values = NULL,
  return_plotly = FALSE,
  marker_method = "presto",
  seurat_assay = "gene",
  hide_technical = NULL,
  unique_markers = FALSE,
  p_val_cutoff = 1,
  ...
)
}
\arguments{
\item{seu}{a seurat object}

\item{metavar}{the metadata variable from which to pick clusters}

\item{num_markers}{default is 5}

\item{return_plotly}{whether to return an interactive ploly plot}

\item{marker_method}{either "presto" or "genesorteR"}

\item{hide_technical}{whether to exclude mitochondrial or ribosomal genes}

\item{...}{}
}
\description{
Plot a dot plot of n marker features grouped by cell metadata
available methods are wilcoxon rank-sum test implemented in
\href{https://github.com/immunogenomics/presto}{presto} and specificity scores implemented in \href{https://github.com/mahmoudibrahim/genesorteR}{genesorteR}
}
\examples{

# interactive mode using "presto"
plot_markers(panc8, metavar = "tech", marker_method = "presto", return_plotly = TRUE)

# static mode using "presto"
plot_markers(panc8, metavar = "tech", marker_method = "genesorteR", return_plotly = FALSE)

}
