% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_markers}
\alias{plot_markers}
\title{Plot Cluster Marker Genes}
\usage{
plot_markers(seu, resolution, num_markers = 5, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot Cluster Marker Genes
}
