<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle3.R
\name{plot_pseudotime}
\alias{plot_pseudotime}
\title{Plot a Monocle Cell Data Set}
\usage{
plot_pseudotime(cds, resolution, color_cells_by = "louvain_cluster")
}
\arguments{
\item{color_cells_by}{}
}
\description{
Plot a Monocle Cell Data Set
}
||||||| merged common ancestors
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{plot_pseudotime}
\alias{plot_pseudotime}
\title{Plot a Monocle Cell Data Set}
\usage{
plot_pseudotime(cds, resolution, color_cells_by = "louvain_cluster")
}
\arguments{
\item{color_cells_by}{}
}
\description{
Plot a Monocle Cell Data Set
}
>>>>>>> 2d43e2725a02764750212f5c9acf866ac6e6a483
