% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{plot_pseudotime}
\alias{plot_pseudotime}
\title{Plot pseudotime on a Monocle Cell Data Set}
\usage{
plot_pseudotime(cds, resolution, color_cells_by = NULL, genes = NULL)
}
\arguments{
\item{color_cells_by}{}
}
\value{

}
\description{
Plot pseudotime on a Monocle Cell Data Set
}
