% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pseudotime_heatmap}
\alias{plot_pseudotime_heatmap}
\title{plot pseudotime heatmap}
\usage{
plot_pseudotime_heatmap(cds_subset, cluster_rows = TRUE,
  hclust_method = "ward.D2", num_clusters = 6, hmcols = NULL,
  add_annotation_row = NULL, add_annotation_col = NULL,
  show_rownames = FALSE, use_gene_short_name = TRUE,
  norm_method = c("log", "vstExprs"), scale_max = 3, scale_min = -3,
  trend_formula = "~sm.ns(Pseudotime, df=3)", return_heatmap = FALSE,
  cores = 1)
}
\arguments{
\item{cores}{}
}
\description{
plot pseudotime heatmap
}
