% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle2.R
\name{plot_pseudotime_heatmap}
\alias{plot_pseudotime_heatmap}
\title{plot pseudotime heatmap}
\usage{
plot_pseudotime_heatmap(
  heatmap_matrix,
  heatmap_title,
  dend_k = 6,
  cluster_rows = T,
  query_set = T,
  hmcols = NULL,
  seriation = F,
  row_font = 4,
  heatmap_height = 96,
  heatmap_width = 8
)
}
\arguments{
\item{heatmap_width}{}
}
\value{

}
\description{
plot pseudotime heatmap
}
