% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_readcount}
\alias{plot_readcount}
\title{Plot Read Count}
\usage{
plot_readcount(
  seu,
  metavar = "nCount_RNA",
  color.by = "batch",
  yscale = "linear"
)
}
\arguments{
\item{yscale}{}
}
\value{

}
\description{
Plot Read Count
}
