% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_readcount}
\alias{plot_readcount}
\title{Plot Read Count}
\usage{
plot_readcount(
  object,
  group_by = NULL,
  fill_by = NULL,
  yscale = "linear",
  return_plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{A object}

\item{group_by}{Metadata variable to plot. Default set to "nCount_RNA"}

\item{fill_by}{Variable to color bins by. Default set to "batch"}

\item{yscale}{Scale of y axis. Default set to "linear"}

\item{return_plotly}{whether to return an interactive plotly plot.}
}
\value{
a histogram of read counts
}
\description{
Draw a box plot for read count data of a metadata variable
}
\examples{

data(small_example_dataset)
small_example_dataset <- object_calcn(small_example_dataset)
# interactive plotly
plot_readcount((small_example_dataset), return_plotly = TRUE)

# static plot
plot_readcount((small_example_dataset), return_plotly = FALSE)
}
