% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_readcount}
\alias{plot_readcount}
\title{Plot Read Count}
\usage{
plot_readcount(
  object,
  group_by = "nCount_RNA",
  color.by = "batch",
  yscale = "linear",
  return_plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{A object}

\item{group_by}{Metadata variable to plot. Default set to "nCount_RNA"}

\item{color.by}{Variable to color bins by. Default set to "batch"}

\item{yscale}{Scale of y axis. Default set to "linear"}

\item{return_plotly}{whether to return an interactive plotly plot. Default set to FALSE}

\item{...}{extra args passed to ggplot2}
}
\value{
a histogram of read counts
}
\description{
Draw a box plot for read count data of a metadata variable
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
# interactive plotly
plot_readcount((chevreul_sce), return_plotly = TRUE)

# static plot
plot_readcount((chevreul_sce), return_plotly = FALSE)
}
