% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{plot_scvelo}
\alias{plot_scvelo}
\title{plot scvelo}
\usage{
plot_scvelo(
  object,
  mode = c("steady_state", "deterministic", "stochastic", "dynamical"),
  ...
)
}
\arguments{
\item{object}{a object}

\item{...}{extra args passed to run_scvelo}

\item{loom_path}{path to matching loom file}

\item{velocity_mode}{deterministic, stochastis, or dynamical}
}
\value{
a single cell object with velocity calculated
}
\description{
run scvelo on a gene or transcript level object
}
