% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{plot_scvelo}
\alias{plot_scvelo}
\title{Plot scvelo on embedding plot}
\usage{
plot_scvelo(
  adata,
  group.by = "batch",
  plot_method = c("stream", "arrow", "dynamics")
)
}
\arguments{
\item{plot_method}{}
}
\description{
Plot scvelo on embedding plot
}
