% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_transcript_composition}
\alias{plot_transcript_composition}
\title{Plot Transcript Composition}
\usage{
plot_transcript_composition(
  seu,
  gene_symbol,
  group.by = "batch",
  standardize = FALSE,
  drop_zero = FALSE
)
}
\arguments{
\item{standardize}{}
}
\description{
plot the proportion of reads of a given gene map to each transcript
}
\examples{
plot_transcript_composition(human_gene_transcript_seu, "RXRG", group.by = "gene_snn_res.0.6")

}
