% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_transcript_composition}
\alias{plot_transcript_composition}
\title{Title}
\usage{
plot_transcript_composition(
  seu_transcript,
  seu_gene,
  gene_symbol,
  group.by = "batch"
)
}
\arguments{
\item{group.by}{}
}
\value{

}
\description{
Title
}
