% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_var}
\alias{plot_var}
\title{Plot Metadata Variables}
\usage{
plot_var(
  seu,
  embedding = "umap",
  group = "batch",
  dims = c(1, 2),
  highlight = NULL,
  pt.size = 1,
  return_plotly = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot Metadata Variables
}
\examples{

plot_var(seurat_pancreas_reduced$gene, group = "batch")
}
