% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_var}
\alias{plot_var}
\title{Plot Metadata Variables}
\usage{
plot_var(
  object,
  group = "batch",
  embedding = "UMAP",
  dims = c(1, 2),
  highlight = NULL,
  pt.size = 1,
  return_plotly = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{group}{Name of one or more metadata columns to group (color) cells by.}

\item{embedding}{The dimensional reduction technique to be used}

\item{dims}{Dimensions to plot, must be a two-length numeric vector}

\item{highlight}{A list of character or numeric vectors of cells to highlight}

\item{pt.size}{Adjust point size on the plot}

\item{return_plotly}{Convert plot to interactive web-based graph}

\item{...}{extra parameters passed to ggplot}
}
\value{
a ggplot
}
\description{
Plots static or interactive plot where each point represents a cell metadata
variable whose position on the map depends on cell embeddings determined by the
reduction technique used
}
\examples{
# static mode
\dontrun{
plot_var(human_gene_transcript_object, group = "batch", return_plotly = FALSE)
}

# interactive plotly plot
\dontrun{
plotly_plot <- plot_var(human_gene_transcript_object, group = "batch")
}
\dontrun{
print(plotly_plot)
}
}
