% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_var}
\alias{plot_var}
\title{Plot Metadata Variables}
\usage{
plot_var(
  seu,
  embedding = "umap",
  group = "batch",
  dims = c(1, 2),
  highlight = NULL,
  pt.size = 1,
  return_plotly = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Plot Metadata Variables
}
\examples{

# static mode
plot_var(panc8, group = "batch", return_plotly  = FALSE)

# interactive plotly plot
plotly_plot <- plot_var(panc8, group = "batch")
print(plotly_plot)

}
