% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_velocity_arrows}
\alias{plot_velocity_arrows}
\title{Plot RNA velocity Computed by Velocyto.R}
\usage{
plot_velocity_arrows(
  seu,
  reduction = "umap",
  format = "arrow",
  arrow.scale = 3
)
}
\arguments{
\item{format}{either an arrow or grid format}

\item{arrow.scale}{}
}
\value{

}
\description{
Plot RNA velocity Computed by Velocyto.R
}
