% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{plot_violin}
\alias{plot_violin}
\title{Make Violin Plots Based on Metadata and Feature Expression}
\usage{
plot_violin(seu, plot_var = "batch", plot_vals = NULL,
  features = "RXRG", ...)
}
\arguments{
\item{seu}{a seurat object}

\item{plot_var}{a metadata value}

\item{plot_vals}{set of values to subset on}

\item{features}{vector of gene or transcript names}

\item{...}{additional arguments to \link[Seurat]{VlnPlot}}
}
\description{
Make Violin Plots Based on Metadata and Feature Expression
}
