% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_violin}
\alias{plot_violin}
\title{plot Violin plot}
\usage{
plot_violin(
  seu,
  plot_var = "batch",
  plot_vals = NULL,
  features = "RXRG",
  assay = "gene",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
plot Violin plot
}
\examples{

plot_violin(panc8, plot_var = "batch", features = c("NRL"))

}
