% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot Dimensionally Reduced Data}
\usage{
plotDimRed(input, output, session, seu, plot_types, featureType, organism_type)
}
\arguments{
\item{organism_type}{}
}
\value{

}
\description{
Plot Dimensionally Reduced Data
}
