% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Plot Dimensional Reduduction}
\usage{
plotDimRed(
  input,
  output,
  session,
  object,
  plot_types,
  featureType,
  organism_type,
  reductions
)
}
\arguments{
\item{reductions}{}
}
\description{
Plot Dimensional Reduduction
}
