% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotVelocity}
\alias{plotVelocity}
\title{RNA Velocity}
\usage{
plotVelocity(input, output, session, seu, loom_path, featureType)
}
\arguments{
\item{featureType}{}
}
\value{

}
\description{
RNA Velocity
}
