% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot Violin Server}
\usage{
plotViolin(input, output, session, seu, featureType, organism_type)
}
\arguments{
\item{input}{one or more feature to be plots}

\item{seu}{Seurat object}

\item{featureType}{Gene or Transcript}

\item{organism_type}{Organism}
}
\description{
Plot Violin Server
}
