% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot Violin server}
\usage{
plotViolin(input, output, session, seu, feature_type)
}
\arguments{
\item{feature_type}{}
}
\value{

}
\description{
Plot Violin server
}
\examples{

}
