% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot Violin Server}
\usage{
plotViolin(input, output, session, seu, featureType, organism_type)
}
\arguments{
\item{organism_type}{}
}
\description{
Plot Violin Server
}
