% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{prep_scvelo}
\alias{prep_scvelo}
\title{scvelo_assay}
\usage{
prep_scvelo(
  seu,
  loom_path,
  plot_method = c("stream", "arrow", "dynamics"),
  ...
)
}
\arguments{
\item{seu}{a seurat object}

\item{loom_path}{path to matching loom file}

\item{plot_method}{plotting method to use from scvelo}

\item{...}{}

\item{group.by}{metadata to color plot}
}
\value{

}
\description{
run scvelo on a gene or transcript level seurat object
}
