% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propagate_spreadsheet_changes}
\alias{propagate_spreadsheet_changes}
\title{Propagate Metadata Changes}
\usage{
propagate_spreadsheet_changes(updated_table, object)
}
\arguments{
\item{updated_table}{updated metadata}

\item{object}{a single cell object}
}
\value{
a single cell object
}
\description{
Propagate Metadata Changes
}
\examples{

\donttest{
updated_table <- read_csv(system.file("extdata", "new_meta.csv", package="chevreul"))

propagate_spreadsheet_changes(updated_table, human_gene_transcript_sce)
}
}
