% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{propagate_spreadsheet_changes}
\alias{propagate_spreadsheet_changes}
\title{Propagate Metadata Changes}
\usage{
propagate_spreadsheet_changes(meta, object)
}
\arguments{
\item{meta}{updated metadata}

\item{object}{a SingleCellExperiment object}
}
\value{
a SingleCellExperiment object
}
\description{
Propagate Metadata Changes
}
\examples{


data(small_example_dataset)
new_meta <- data.frame(row.names = colnames(small_example_dataset))
new_meta$example <- "example"

propagate_spreadsheet_changes(new_meta, small_example_dataset)
}
