% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classes.R
\name{pull_metadata}
\alias{pull_metadata}
\title{Pull object metadata}
\usage{
pull_metadata(object)
}
\arguments{
\item{object}{a single cell object}
}
\value{
object metadata
}
\description{
Pull object metadata
}
