% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{query_experiment}
\alias{query_experiment}
\title{Query Assay}
\usage{
query_experiment(object, experiment)
}
\arguments{
\item{object}{a single cell object}

\item{experiment}{an experiment name}
}
\value{
logical scalar indicating if experiment is present in object
}
\description{
Query Assay
}
\examples{
query_experiment(human_gene_transcript_sce, "gene")
}
