% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{query_experiment}
\alias{query_experiment}
\title{Query Experiment}
\usage{
query_experiment(object, experiment)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{experiment}{an experiment name}
}
\value{
logical scalar indicating if experiment is present in object
}
\description{
Query Experiment
}
