% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_project_db}
\alias{read_project_db}
\title{Read a database of chevreul projects}
\usage{
read_project_db(
  projects_dir = NULL,
  cache_location = "~/.cache/chevreul",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Read a database of chevreul projects
}
