% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_project_db}
\alias{read_project_db}
\title{Read a database of chevreul projects}
\usage{
read_project_db(
  cache_location = "~/.cache/chevreul",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{cache_location}{Path to cache "~/.cache/chevreul"}

\item{sqlite_db}{sqlite db}

\item{verbose}{print messages}
}
\value{
a tibble with SingleCellExperiment objects
}
\description{
Reads database of chevreul projects to a data frame
}
