% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{record_experiment_data}
\alias{record_experiment_data}
\title{Record Experiment Metadata}
\usage{
record_experiment_data(
  object,
  experiment_name = "default_experiment",
  organism = "human"
)
}
\arguments{
\item{object}{A object}

\item{organism}{}
}
\value{
a single cell object
}
\description{
Records miscellaneous data
}
