% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{record_experiment_data}
\alias{record_experiment_data}
\title{Record Experiment Metadata}
\usage{
record_experiment_data(object, experiment_name = NULL, organism = NULL)
}
\arguments{
\item{add_all_meta_data}{}
}
\value{

}
\description{
Record Experiment Metadata
}
