% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{record_experiment_data}
\alias{record_experiment_data}
\title{Record Experiment Metadata}
\usage{
record_experiment_data(
  object,
  experiment_name = "default_experiment",
  organism = "human"
)
}
\arguments{
\item{object}{A object}

\item{experiment_name}{name of the experiment}

\item{organism}{human or mouse}
}
\value{
a SingleCellExperiment object
}
\description{
Records miscellaneous data
}
\examples{

data(small_example_dataset)
record_experiment_data(small_example_dataset)

}
