% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_edit.R
\name{reformatMetadataDR}
\alias{reformatMetadataDR}
\title{Reformat Seurat Object Metadata Server}
\usage{
reformatMetadataDR(
  input,
  output,
  session,
  seu,
  featureType = "gene",
  col_bind = NULL,
  col_edit = TRUE,
  col_options = NULL,
  col_stretch = FALSE,
  col_names = TRUE,
  col_readonly = NULL,
  col_factor = FALSE,
  row_bind = NULL,
  row_edit = TRUE,
  quiet = FALSE,
  hide = FALSE
)
}
\arguments{
\item{seu}{}
}
\value{

}
\description{
Reformat Seurat Object Metadata Server
}
