% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_features.R
\name{regress_by_features}
\alias{regress_by_features}
\title{Regress SingleCellExperiment Object by Given Set of Genes}
\usage{
regress_by_features(object, feature_set, set_name, regress = TRUE)
}
\arguments{
\item{object}{A object}

\item{feature_set}{a set of features}

\item{set_name}{as a string}

\item{regress}{whether to regress}
}
\value{
a single cell object with features regressed
}
\description{
Regress SingleCellExperiment Object by Given Set of Genes
}
\examples{

regressed_object <- regress_by_features(panc8, feature_set = cc.genes$s.genes, set_name = "s_genes")

}
