% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_features.R
\name{regress_by_features}
\alias{regress_by_features}
\title{Regress Seurat Object by Given Set of Genes}
\usage{
regress_by_features(seu, feature_set, set_name, regress = TRUE, ...)
}
\arguments{
\item{set_name}{as a string}

\item{...}{}

\item{gene_set}{as a length 1 list containing a character vector of gene symbols}
}
\description{
Regress Seurat Object by Given Set of Genes
}
\examples{

regressed_seu <- regress_by_features(panc8, feature_set  = cc.genes$s.genes, set_name = "s_genes")

}
