% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_features.R
\name{regress_cell_cycle}
\alias{regress_cell_cycle}
\title{Regress SingleCellExperiment Object by Given Set of Genes}
\usage{
regress_cell_cycle(object, regress = TRUE)
}
\arguments{
\item{object}{A object}

\item{regress}{whether to regress}
}
\value{
a SingleCellExperiment object with features regressed
}
\description{
Regress SingleCellExperiment Object by Given Set of Genes
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()

regressed_object <- regress_cell_cycle(chevreul_sce)

}
