% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_features.R
\name{regress_cell_cycle}
\alias{regress_cell_cycle}
\title{Regress SingleCellExperiment Object by Given Set of Genes}
\usage{
regress_cell_cycle(object)
}
\arguments{
\item{object}{A object}
}
\value{
a SingleCellExperiment object with features regressed
}
\description{
Regress SingleCellExperiment Object by Given Set of Genes
}
