% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{reintegrate_object}
\alias{reintegrate_object}
\title{Reintegrate (filtered) singlecell objectss}
\usage{
reintegrate_object(
  object,
  feature = "gene",
  suffix = "",
  reduction = "pca",
  algorithm = 1,
  ...
)
}
\arguments{
\item{object}{A singlecell objects}

\item{suffix}{to be appended to file saved in output dir}

\item{reduction}{to use default is pca}
}
\value{
a single cell object
}
\description{
\enumerate{
\item split by batch
\item integrate
\item run integration pipeline and save
}
}
