% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{reintegrate_seu}
\alias{reintegrate_seu}
\title{Reintegrate (filtered) seurat objects}
\usage{
reintegrate_seu(seu, feature = "gene", suffix = "", reduction = "pca", ...)
}
\arguments{
\item{suffix}{to be appended to file saved in output dir}

\item{reduction}{to use default is pca}
}
\value{

}
\description{
\enumerate{
\item split by batch
\item integrate
\item run integration pipeline and save
}
}
