% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{rename_object}
\alias{rename_object}
\title{Give a new project name to a single cell object}
\usage{
rename_object(object, new_name)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{new_name}{New name to assign}
}
\value{
a renamed single cell object
}
\description{
Give a new project name to a single cell object
}
\examples{
chevreul_sce <- chevreuldata::human_gene_transcript_sce()
rename_object(chevreul_sce, "new_name")
}
