% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{rename_seurat}
\alias{rename_seurat}
\title{Give a new project name to a seurat object}
\usage{
rename_seurat(seu, new_name)
}
\arguments{
\item{new_name}{}
}
\value{

}
\description{
Give a new project name to a seurat object
}
