% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{retrieve_experiment}
\alias{retrieve_experiment}
\title{Retrieve Assay}
\usage{
retrieve_experiment(object, experiment)
}
\arguments{
\item{experiment}{}
}
\description{
Retrieve Assay
}
\examples{
retrieve_experiment(human_gene_transcript_sce)
}
