% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{retrieve_experiment}
\alias{retrieve_experiment}
\title{Retrieve Assay}
\usage{
retrieve_experiment(object, experiment)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{experiment}{an experiment name}
}
\value{
Main or alt experiment in an object
}
\description{
Retrieve Assay
}
\examples{

data(small_example_dataset)
mainExpName(small_example_dataset) <- "gene"
retrieve_experiment(small_example_dataset, experiment = "gene")
}
