% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{run_enrichmentbrowser}
\alias{run_enrichmentbrowser}
\title{Run Enrichment Browser on Differentially Expressed Genes}
\usage{
run_enrichmentbrowser(
  seu,
  cluster_list,
  de_results,
  enrichment_method = c("ora"),
  ...
)
}
\arguments{
\item{de_results}{}
}
\value{

}
\description{
Run Enrichment Browser on Differentially Expressed Genes
}
