% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{run_object_de}
\alias{run_object_de}
\title{Run Seurat Differential Expression}
\usage{
run_object_de(
  object,
  cluster1,
  cluster2,
  resolution,
  diffex_scheme = "louvain",
  featureType,
  tests = c("t", "wilcox", "bimod")
)
}
\arguments{
\item{diffex_scheme}{}
}
\description{
Run Seurat Differential Expression
}
