% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{run_scvelo}
\alias{run_scvelo}
\title{scvelo_assay}
\usage{
run_scvelo(
  object,
  loom_path,
  assay = "gene",
  fit.quantile = 0.05,
  check_loom = FALSE
)
}
\arguments{
\item{object}{a object}

\item{loom_path}{path to matching loom file}

\item{assay}{gene}

\item{fit.quantile}{how to fit velocity}

\item{check_loom}{FALSE}
}
\value{
a single cell object with RNA velocity calculated
}
\description{
run scvelo on a gene or transcript level object
}
