% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{save_object}
\alias{save_object}
\title{Save object to \if{html}{\out{<project>}}/output/sce/\if{html}{\out{<feature>}}_object.rds}
\usage{
save_object(object, prefix = "unfiltered", proj_dir = getwd())
}
\arguments{
\item{proj_dir}{}

\item{...}{named arguments specifying objects list of objects; default "gene" and "transcript"}
}
\description{
Save object to \if{html}{\out{<project>}}/output/sce/\if{html}{\out{<feature>}}_object.rds
}
\examples{
\dontrun{
save_object(gene = feature_objects$gene, transcript = feature_objects$transcript, proj_dir = proj_dir)

save_object(gene = feature_objects$gene, transcript = feature_objects$transcript, prefix = "remove_nonPRs", proj_dir = proj_dir)
}
}
