% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{save_seurat}
\alias{save_seurat}
\title{Save seurat object to <project>/output/sce/<feature>_seu.rds}
\usage{
save_seurat(seu, feature, suffix = "", temp = F)
}
\arguments{
\item{feature}{}
}
\description{
Save seurat object to <project>/output/sce/<feature>_seu.rds
}
