% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{set_cell_metadata}
\alias{set_cell_metadata}
\title{Set cell metadata from a given object}
\usage{
set_cell_metadata(object, meta)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{meta}{a dataframe containing object metadata}
}
\value{
a SingleCellExperiment object with new colData
}
\description{
Set cell metadata from a given object
}
\examples{


chevreul_sce <- mockSCE(ncells=200, ngenes=1000)
new_meta <- data.frame(row.names = colnames(chevreul_sce))
new_meta$example <- "example"
set_cell_metadata(chevreul_sce, new_meta)
}
