% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{set_feature_type}
\alias{set_feature_type}
\title{Set Feature Types}
\usage{
set_feature_type(object, feature_type)
}
\arguments{
\item{object}{a SingleCellExperiment object}

\item{feature_type}{feature type}
}
\value{
an object with assigned feature type
}
\description{
Set Feature Types
}
\examples{


chevreul_sce <- mockSCE(ncells=200, ngenes=1000)
set_feature_type(chevreul_sce, "transcript")
}
