% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{SetDefaultAssay}
\alias{SetDefaultAssay}
\title{Reset default assay}
\usage{
SetDefaultAssay(seu, new_assay)
}
\arguments{
\item{seu}{A seurat object}

\item{new_assay}{Assay to set as default assay}
}
\description{
FUnction to reset the default assay of a seurat object
}
