% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{SetDefaultAssay}
\alias{SetDefaultAssay}
\title{Reset the default assay of a seurat object}
\usage{
SetDefaultAssay(seu, new_assay)
}
\arguments{
\item{new_assay}{}
}
\description{
Reset the default assay of a seurat object
}
