% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{seu_complex_heatmap}
\alias{seu_complex_heatmap}
\title{Plot Annotated Complexheatmap from Seurat object}
\usage{
seu_complex_heatmap(
  seu,
  features = NULL,
  group.by = "ident",
  cells = NULL,
  slot = "scale.data",
  assay = NULL,
  group.bar.height = 0.01,
  column_split = NULL,
  col_arrangement = "ward.D2",
  mm_col_dend = 30,
  ...
)
}
\arguments{
\item{features}{Vector of features to plot. Features can come}

\item{group.by}{Name of one or more metadata columns to annotate columns by (for example, orig.ident)}

\item{cells}{Cells to retain}

\item{column_split}{whether to split columns by metadat value}

\item{col_arrangement}{how to arrange columns whether with a dendrogram (Ward.D2, average, etc.) or exclusively by metadata category}

\item{mm_col_dend}{height of column dendrogram}

\item{...}{additional arguments passed to ComplexHeatmap::Heatmap}
}
\description{
Plot Annotated Complexheatmap from Seurat object
}
\examples{

# plot top 50 variable genes
top_50_features <- VariableFeatures(panc8)[1:50]
seu_complex_heatmap(panc8, features = top_50_features)

}
