% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{seu_complex_heatmap}
\alias{seu_complex_heatmap}
\title{Plot Annotated Complexheatmap from Seurat object}
\usage{
seu_complex_heatmap(
  object,
  features = NULL,
  cells = NULL,
  group.by = "ident",
  slot = "scale.data",
  assay = NULL,
  group.bar.height = 0.01,
  cluster_columns = FALSE,
  column_split = NULL,
  col_dendrogram = "ward.D2",
  mm_col_dend = 30,
  ...
)
}
\arguments{
\item{cluster_columns}{FALSE}

\item{...}{}
}
\value{

}
\description{
Plot Annotated Complexheatmap from Seurat object
}
