% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{seu_complex_heatmap}
\alias{seu_complex_heatmap}
\title{Plot Annotated Complexheatmap from Seurat object}
\usage{
seu_complex_heatmap(
  seu,
  features = NULL,
  group.by = "ident",
  cells = NULL,
  slot = "scale.data",
  assay = NULL,
  group.bar.height = 0.01,
  cluster_columns = FALSE,
  column_split = NULL,
  col_dendrogram = "ward.D2",
  mm_col_dend = 30,
  ...
)
}
\arguments{
\item{cluster_columns}{FALSE}

\item{...}{}
}
\description{
Plot Annotated Complexheatmap from Seurat object
}
\examples{

# plot top 50 variable genes
top_50_features <- VariableFeatures(panc8)[1:50]
seu_complex_heatmap(panc8, features = top_50_features)

}
