% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{seu_from_tibbles}
\alias{seu_from_tibbles}
\title{Create a Seurat Object from a set of tibbles}
\usage{
seu_from_tibbles(exp_tbl, feature, meta_tbl, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Create a Seurat Object from a set of tibbles
}
