% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seu.R
\name{seu_from_tximport}
\alias{seu_from_tximport}
\title{Create a Seurat Object from return of tximport}
\usage{
seu_from_tximport(txi, meta_tbl, ...)
}
\arguments{
\item{...}{}
}
\description{
Create a Seurat Object from return of tximport
}
