% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{seurat_integrate}
\alias{seurat_integrate}
\title{Batch Correct Multiple Single Cell Objects}
\usage{
seurat_integrate(object_list, method = "cca", organism = "human", ...)
}
\arguments{
\item{object_list}{List of two or more single cell objects}

\item{method}{Default "cca"}

\item{...}{}
}
\value{
an integrated single cell object
}
\description{
Batch Correct Multiple Single Cell Objects
}
