% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{seurat_integration_pipeline}
\alias{seurat_integration_pipeline}
\title{Run Seurat Integration}
\usage{
seurat_integration_pipeline(
  seu_list,
  resolution = seq(0.2, 2, by = 0.2),
  suffix = "",
  algorithm = 1,
  organism = "human",
  annotate_cell_cycle = FALSE,
  annotate_percent_mito = FALSE,
  reduction = "pca",
  ...
)
}
\arguments{
\item{suffix}{a suffix to be appended to a file save in output dir}

\item{...}{}
}
\description{
run batch correction, followed by:
\enumerate{
\item stashing of batches in metadata 'batch'
\item clustering with resolution 0.2 to 2.0 in increments of 0.2
\item saving to <proj_dir>/output/sce/<feature>\emph{seu}<suffix>.rds
}
}
\examples{

batches <- panc8 \%>\%
  Seurat::SplitObject(split.by = "tech")

integrated_seu <- seurat_integration_pipeline(batches)
}
