% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{seurat_integration_pipeline}
\alias{seurat_integration_pipeline}
\title{Run Seurat Integration}
\usage{
seurat_integration_pipeline(seus, resolution = seq(0.2, 2, by = 0.2),
  suffix = "", ...)
}
\arguments{
\item{suffix}{a suffix to be appended to a file save in output dir}

\item{...}{}
}
\description{
run batch correction, followed by:
\enumerate{
\item stashing of batches in metadata 'batch'
\item clustering with resolution 0.2 to 2.0 in increments of 0.2
\item saving to <proj_dir>/output/sce/<feature>\emph{seu}<suffix>.rds
}
}
