% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{seurat_pipeline}
\alias{seurat_pipeline}
\title{Run Seurat Pipeline}
\usage{
seurat_pipeline(
  seu,
  assay = "gene",
  resolution = 0.6,
  reduction = "pca",
  organism = "human",
  ...
)
}
\arguments{
\item{organism}{}
}
\description{
Preprocess, Cluster and Reduce Dimensions for a single seurat object
}
\examples{

processed_seu <- seurat_pipeline(panc8)

}
