% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{seurat_pipeline}
\alias{seurat_pipeline}
\title{Run Seurat Pipeline}
\usage{
seurat_pipeline(
  seu,
  feature = "gene",
  resolution = 0.6,
  reduction = "pca",
  annotate_cell_cycle = TRUE,
  annotate_percent_mito = TRUE,
  select_var_features = TRUE,
  ...
)
}
\arguments{
\item{resolution}{}
}
\value{

}
\description{
Preprocess, Cluster and Reduce Dimensions for a single seurat object
}
