% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{seurat_pipeline}
\alias{seurat_pipeline}
\title{Run Seurat Pipeline}
\usage{
seurat_pipeline(seu = seu, resolution = 0.6, reduction = "pca", ...)
}
\arguments{
\item{resolution}{}
}
\description{
Preprocess, Cluster and Reduce Dimensions for a single seurat object
}
