% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{seurat_preprocess}
\alias{seurat_preprocess}
\title{Preprocess Seurat Object}
\usage{
seurat_preprocess(seu, scale = TRUE, normalize = TRUE, ...)
}
\arguments{
\item{scale}{}
}
\description{
Preprocess Seurat Object
}
