% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{seurat_reduce_dimensions}
\alias{seurat_reduce_dimensions}
\title{Dimensional Reduction}
\usage{
seurat_reduce_dimensions(seu, reduction = "pca", ...)
}
\arguments{
\item{seu}{}
}
\value{

}
\description{
Run PCA, TSNE and UMAP on a seurat object
perplexity should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation
}
