% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_batch_correct.R
\name{seurat_reduce_dimensions}
\alias{seurat_reduce_dimensions}
\title{Dimensional Reduction}
\usage{
seurat_reduce_dimensions(seu)
}
\arguments{
\item{seu}{}
}
\description{
Run PCA, TSNE and UMAP on a seurat object
}
