% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{seuratApp}
\alias{seuratApp}
\title{Create a Shiny App to View Seurat Data}
\usage{
seuratApp(proj_dir, plotTypes, filterTypes, appTitle)
}
\arguments{
\item{proj_dir}{The base directory of the project ex. "~/single_cell_project/project/"}

\item{plotTypes}{The variables to use for the Embedding plots, ex. c("batch", "treatment_group")}

\item{filterTypes}{a named vector of file suffixes corresponding to filtered seurat objects in the project ex. c("Without Low Read Count" = "wo_lowrc")}

\item{appTitle}{The Title of the App}
}
\description{
Create a Shiny App to View Seurat Data
}
