% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{seuratApp}
\alias{seuratApp}
\title{Create Seurat App}
\usage{
seuratApp(preset_project, filterTypes, appTitle, feature_types = "gene",
  futureMb = 849)
}
\arguments{
\item{preset_project}{A preloaded project to start the app with}

\item{filterTypes}{A named vector of file suffixes corresponding to subsets of the data}

\item{appTitle}{A title of the App}

\item{futureMb}{amount of Mb allocated to future package}

\item{preset_project}{}
}
\description{
Create Seurat App
}
