% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{seuratApp}
\alias{seuratApp}
\title{Create Seurat App}
\usage{
seuratApp(proj_dir, plot_types, filterTypes, appTitle, futureMb = 849,
  ...)
}
\arguments{
\item{proj_dir}{The project directory of the base dataset ex. "~/single_cell_projects/sc_cone_devel/proj"}

\item{plot_types}{The types of plots to be shown as a named list containing two vectors: 1) category_vars and 2) continuous_vars}

\item{filterTypes}{A named vector of file suffixes corresponding to subsets of the data, ex. filterTypes <- c("", "remove_lowrc") %>% set_names(c("Unfiltered", "low read count cells"))}

\item{appTitle}{A title of the App}

\item{futureMb}{amount of Mb allocated to future package}

\item{...}{}
}
\description{
Creat a shiny app to view and manipulate a Seurat object
}
