% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_app.R
\name{seuratApp}
\alias{seuratApp}
\title{Create Seurat App}
\usage{
seuratApp(
  preset_project,
  appTitle = "seuratTools",
  organism_type = "human",
  db_path = "~/.cache/seuratTools/single-cell-projects.db",
  futureMb = 13000
)
}
\arguments{
\item{preset_project}{A preloaded project to start the app with}

\item{appTitle}{A title of the App}

\item{organism_type}{}

\item{futureMb}{amount of Mb allocated to future package}
}
\value{

}
\description{
Create Seurat App
}
