% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{splitByCol}
\alias{splitByCol}
\title{Split SingleCellExperiment by colData variable}
\usage{
splitByCol(x, f = "batch")
}
\arguments{
\item{x}{SingleCellExperiment object}

\item{f}{colData variable as a string}
}
\value{
a list of singlecellexperiments name by colData value
}
\description{
Split SingleCellExperiment by colData variable
}
\examples{


chevreul_sce <- mockSCE(ncells=200, ngenes=1000)
splitByCol(chevreul_sce, "batch")
}
