% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{stash_marker_features}
\alias{stash_marker_features}
\title{Stash Marker Genes in a Seurat Object}
\usage{
stash_marker_features(
  metavar,
  seu,
  seurat_assay,
  top_n = 200,
  p_val_cutoff = 0.5
)
}
\arguments{
\item{top_n}{}
}
\description{
Marker Genes will be stored in slot \verb{@misc$markers}
}
\examples{

seu <- stash_marker_features(metavar = "batch", seu, seurat_assay = "gene")

}
