% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cluster.R
\name{stash_marker_features}
\alias{stash_marker_features}
\title{Stash Marker Genes in a Seurat Object}
\usage{
stash_marker_features(metavar, seu, top_n = 200)
}
\arguments{
\item{top_n}{}
}
\value{

}
\description{
Marker Genes will be stored in slot \verb{@misc$markers}
}
