% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{stash_marker_features}
\alias{stash_marker_features}
\title{Stash Marker Genes in a SingleCellExperiment Object}
\usage{
stash_marker_features(
  object,
  group_by,
  experiment = "gene",
  top_n = 200,
  p_val_cutoff = 0.5
)
}
\arguments{
\item{object}{A object}

\item{group_by}{A metadata variable to group by}

\item{experiment}{An experiment to use}

\item{top_n}{Use top n genes, Default 200}

\item{p_val_cutoff}{p value cut-off, Default value is "0.5"}
}
\value{
a SingleCellExperiment object with marker genes
}
\description{
Marker Genes will be stored in object metadata as \code{markers}
}
