% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_annotation.R
\name{transcripts_to_genes}
\alias{transcripts_to_genes}
\title{Ensembl Transcript Ids to Gene Symbols}
\usage{
transcripts_to_genes(transcripts, organism = "human")
}
\arguments{
\item{transcripts}{}
}
\value{
a vector of gene symbols
}
\description{
convert ensembl transcript ids to hgnc gene symbols
}
\examples{

RXRG_transcripts_hs <- c("ENST00000359842", "ENST00000470566", "ENST00000465764", "ENST00000619224")

transcripts_to_genes(transcripts = RXRG_transcripts_hs)

}
