% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{transcripts_to_genes}
\alias{transcripts_to_genes}
\title{Ensembl Transcript Ids to Gene Symbols}
\usage{
transcripts_to_genes(transcripts, organism = "human")
}
\arguments{
\item{transcripts}{transcripts}

\item{organism}{human or mouse}
}
\value{
a vector of gene symbols
}
\description{
Convert ensembl transcript ids to hgnc gene symbols
}
\examples{

NRL_transcripts_hs <-
c("ENST00000359842", "ENST00000470566", "ENST00000465764")

 data("grch38_tx2gene")
 data("grch38")
transcripts_to_genes(transcripts = NRL_transcripts_hs)

}
