% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{unite_metadata}
\alias{unite_metadata}
\title{Unite metadata}
\usage{
unite_metadata(object, group_bys)
}
\arguments{
\item{object}{A object}

\item{group_bys}{A feature or variable to combine}
}
\value{
an object with Idents formed from concatenation of group_bys
}
\description{
Unite metadata
}
\examples{
unite_metadata(human_gene_transcript_sce, "nFeature_gene")

}
