% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{unite_metadata}
\alias{unite_metadata}
\title{Unite metadata}
\usage{
unite_metadata(object, metavars)
}
\arguments{
\item{object}{A object}

\item{metavars}{A feature or variable to combine}
}
\value{
an object with Idents formed from concatenation of metavars
}
\description{
Unite metadata
}
