% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_child_exp_metadata.R
\name{update_exp_meta}
\alias{update_exp_meta}
\title{Update experiment Metadata}
\usage{
update_exp_meta(original_meta, corrected_meta)
}
\arguments{
\item{corrected_meta}{}
}
\description{
Update experiment Metadata
}
