% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{update_human_gene_symbols}
\alias{update_human_gene_symbols}
\title{Update human gene symbols in seurat object}
\usage{
update_human_gene_symbols(seu, assay = "RNA")
}
\arguments{
\item{seu}{}
}
\value{

}
\description{
Update human gene symbols in seurat object
}
