% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_species.R
\name{update_human_gene_symbols}
\alias{update_human_gene_symbols}
\title{Update human gene symbols in object}
\usage{
update_human_gene_symbols(object, assay = "gene")
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{assay}{Assay to use, Default = "gene"}
}
\value{
a single cell object
}
\description{
Update human gene symbols in object
}
