% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_project_db}
\alias{update_project_db}
\title{Update a database of chevreul projects}
\usage{
update_project_db(
  projects_dir = NULL,
  cache_location = "~/.cache/chevreul",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{projects_dir}{The project directory to be updated}

\item{cache_location}{Path to cache "~/.cache/chevreul"}

\item{sqlite_db}{sqlite db}

\item{verbose}{print messages}
}
\value{
a sqlite database with SingleCellExperiment objects
}
\description{
Add new/update existing projects to the database by recursing fully
}
