% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seurat_from_child_metadata.R
\name{update_seu_meta}
\alias{update_seu_meta}
\title{Update Seurat Metadata from Project}
\usage{
update_seu_meta(seu, proj_dir, numcols)
}
\arguments{
\item{numcols}{}
}
\value{

}
\description{
Given a project metadata file located in <proj_dir>/data/<meta_file>, update an existing seurat object with the project data
}
