% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_velocity.R
\name{velocyto_assay}
\alias{velocyto_assay}
\title{run velocyto on a gene or transcript level seurat object}
\usage{
velocyto_assay(seu, loom_paths, fit.quantile = 0.05)
}
\arguments{
\item{seu}{a seurat object}

\item{loom_paths}{one or more paths to loom files as a list}

\item{fit.quantile}{}
}
\value{

}
\description{
run velocyto on a gene or transcript level seurat object
}
