% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocyto.R
\name{velocyto_assay}
\alias{velocyto_assay}
\title{run velocyto on a gene or transcript level seurat object}
\usage{
velocyto_assay(seu, loom_path, fit.quantile = 0.05, ...)
}
\arguments{
\item{seu}{a seurat object}

\item{loom_path}{path to matching loom file}

\item{...}{}
}
\value{

}
\description{
run velocyto on a gene or transcript level seurat object
}
